refactor: modularize map-listings component (4,844 → 1,393 lines)
Restructure monolithic component into organized folders:
- components/ (9 UI components + filters)
- helper-functions/ (formatters, auto-center, map-utils)
- types/ (17 interfaces)

Improves maintainability and testability with no breaking changes.
