+++ src/components/map-listings/map-listings.tsx
@@ -1,3396 +1,35 @@
 import React, { useRef, useEffect, useState, useCallback } from "react";
 import mapboxgl from "mapbox-gl";
 import "mapbox-gl/dist/mapbox-gl.css";
-import { BedDouble, Bath, Car, ChevronDown, X } from "lucide-react";
 
-export interface MapListingsProps {
-  /** Repliers API key - required */
-  apiKey: string;
-  /** MapBox access token - required */
-  mapboxToken: string;
-  /** Initial map center coordinates [lng, lat] - auto-detects if not provided */
-  initialCenter?: [number, number];
-  /** Initial zoom level - auto-calculates if not provided */
-  initialZoom?: number;
-  /** Map container height */
-  height?: string;
-  /** Map container width */
-  width?: string;
-  /** Map style */
-  mapStyle?: string;
-  /** Center calculation method */
-  centerCalculation?: "average" | "city";
-  /** Show property count display */
-  showPropertyCount?: boolean;
-}
+// Type imports
+import type {
+  MapListingsProps,
+  ClusterFeature,
+  PropertyFeature,
+  Coordinate,
+  MapFilters,
+  ListingResult,
+} from "./types";
+
+// Component imports
+import { PropertyTooltip } from "./components/PropertyTooltip";
+import { ClusterTooltip } from "./components/ClusterTooltip";
+import { FilterPanel } from "./components/filters/FilterPanel";
+
+// Utility imports
+import { createPriceBubble } from "./helper-functions/map-utils";
+import {
+  detectAutoCenter,
+  getCachedAutoCenter,
+  extractCoordinates,
+  getPolygonBounds,
+  calculateAverageCenter,
+} from "./helper-functions/auto-center";
+
+// Re-export types for external use
+export type { MapListingsProps };
 
-interface ClusterFeature {
-  type: "Feature";
-  properties: {
-    count: number;
-    precision: number;
-    id: string;
-    members?: any[]; // Store actual cluster members from API
-    apiCount?: number; // Original unfiltered count from API for debugging
-  };
-  geometry: {
-    type: "Point";
-    coordinates: [number, number];
-  };
-}
-
-interface PropertyFeature {
-  type: "Feature";
-  properties: {
-    mlsNumber: string;
-    listPrice?: number;
-    propertyType?: string; // "Sale" or "Lease"
-    isProperty: true;
-    // Enhanced fields available at zoom 13+
-    address?: {
-      city?: string;
-      streetDirection?: string;
-      streetName?: string;
-      streetNumber?: string;
-      streetSuffix?: string;
-      state?: string;
-    };
-    details?: {
-      numBedrooms?: number;
-      numBedroomsPlus?: number;
-      numBathrooms?: number;
-      numBathroomsPlus?: number;
-      numGarageSpaces?: number;
-      propertyType?: string;
-    };
-    images?: Array<string>;
-    type?: string;
-    lastStatus?: string;
-    status?: string;
-    class?: string;
-  };
-  geometry: {
-    type: "Point";
-    coordinates: [number, number];
-  };
-}
-
-interface Coordinate {
-  lat: number;
-  lng: number;
-}
-
-interface MapBounds {
-  north: number;
-  south: number;
-  east: number;
-  west: number;
-}
-
-interface AutoCenterData {
-  center: [number, number];
-}
-
-interface MapFilters {
-  listingType: "sale" | "lease";
-  propertyTypes: string[];
-  minPrice?: number;
-  maxPrice?: number | null; // null represents "Max" (no limit)
-  bedrooms?: "all" | "0" | "1" | "2" | "3" | "4" | "5+";
-  bathrooms?: "all" | "1+" | "2+" | "3+" | "4+" | "5+";
-  garageSpaces?: "all" | "1+" | "2+" | "3+" | "4+" | "5+";
-  minSqft?: number;
-  maxSqft?: number | null; // null represents "Max" (no limit)
-  openHouse?: "all" | "today" | "thisWeek" | "thisWeekend" | "anytime";
-  maxMaintenanceFee?: number | null; // null represents "Max" (no limit)
-  activeListingDays?: "1" | "3" | "7" | "30" | "90" | "all" | "15+" | "30+" | "60+" | "90+"; // Active listing date filter
-  soldListingDays?: "1" | "3" | "7" | "30" | "90" | "180" | "360" | "2025" | "2024" | "2023" | "2022" | "2021" | "2020" | "2019" | "2018" | "2017" | "2016" | "2015" | "2014" | "2013" | "2012" | "2011" | "2010" | "2009" | "2008" | "2007"; // Sold listing date filter
-  unavailableListingDays?: "1" | "3" | "7" | "30" | "90" | "180" | "360" | "2025" | "2024" | "2023" | "2022" | "2021" | "2020" | "2019" | "2018" | "2017" | "2016" | "2015" | "2014" | "2013" | "2012" | "2011" | "2010" | "2009" | "2008" | "2007"; // Unavailable listing date filter
-}
-
-interface ListingResult {
-  mlsNumber: string;
-  listPrice: number;
-  address?: {
-    city?: string;
-    streetDirection?: string;
-    streetName?: string;
-    streetNumber?: string;
-    streetSuffix?: string;
-    state?: string;
-  };
-  details?: {
-    numBedrooms?: number;
-    numBedroomsPlus?: number;
-    numBathrooms?: number;
-    numBathroomsPlus?: number;
-    numGarageSpaces?: number;
-    propertyType?: string;
-  };
-  images?: Array<string>;
-  type?: string;
-  lastStatus?: string;
-  status?: string;
-}
-
-interface ListingPreviewProps {
-  listing: ListingResult;
-  onClick?: () => void;
-}
-
-// Shared formatting utilities
-const formatPrice = (price: number) => {
-  return new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency: "USD",
-    minimumFractionDigits: 0,
-    maximumFractionDigits: 0,
-  }).format(price);
-};
-
-const formatBedrooms = (details: ListingResult["details"]) => {
-  if (!details) return "N/A";
-
-  const total = details.numBedrooms;
-  const plus = details.numBedroomsPlus;
-
-  if (total && plus && plus > 0) {
-    return `${total} + ${plus}`;
-  }
-  return total?.toString() || "N/A";
-};
-
-const formatBathrooms = (details: ListingResult["details"]) => {
-  if (!details) return "N/A";
-
-  const total = details.numBathrooms;
-  const plus = details.numBathroomsPlus;
-
-  if (total && plus && plus > 0) {
-    return `${total} + ${plus}`;
-  }
-  return total?.toString() || "N/A";
-};
-
-const formatParking = (details: ListingResult["details"]) => {
-  if (!details) return "N/A";
-  return details.numGarageSpaces?.toString() || "N/A";
-};
-
-const getStatusLabel = (
-  type?: string,
-  lastStatus?: string
-): string | null => {
-  if (lastStatus === "New" || lastStatus === "Pc" || lastStatus === "Ext") {
-    if (type === "Sale") return "For Sale";
-    if (type === "Lease") return "For Lease";
-    return type || null;
-  }
-
-  if (!lastStatus) return null;
-
-  const lastStatusMap: Record<string, string> = {
-    Sus: "Suspended",
-    Exp: "Expired",
-    Sld: "Sold",
-    Ter: "Terminated",
-    Dft: "Deal Fell Through",
-    Lsd: "Leased",
-    Sc: "Sold Conditionally",
-    Sce: "Sold Conditionally (Escape Clause)",
-    Lc: "Leased Conditionally",
-    Cs: "Coming Soon",
-  };
-
-  return lastStatusMap[lastStatus] || lastStatus;
-};
-
-const StatusTag = ({
-  type,
-  lastStatus,
-  size = "sm",
-}: {
-  type?: string;
-  lastStatus?: string;
-  size?: "xs" | "sm";
-}) => {
-  const label = getStatusLabel(type, lastStatus);
-  if (!label) return null;
-
-  let style = "bg-green-100 text-green-600"; // Default for active listings
-
-  switch (label) {
-    case "For Sale":
-      style = "bg-green-100 text-green-600";
-      break;
-    case "For Lease":
-      style = "bg-purple-100 text-purple-600";
-      break;
-    case "Sold":
-      style = "bg-blue-100 text-blue-600";
-      break;
-    case "Leased":
-      style = "bg-purple-100 text-purple-600";
-      break;
-    case "Sold Conditionally":
-    case "Sold Conditionally (Escape Clause)":
-    case "Leased Conditionally":
-      style = "bg-amber-100 text-amber-600";
-      break;
-    case "Coming Soon":
-      style = "bg-indigo-100 text-indigo-600";
-      break;
-    case "Suspended":
-      style = "bg-orange-100 text-orange-600";
-      break;
-    case "Expired":
-      style = "bg-gray-100 text-gray-500";
-      break;
-    case "Terminated":
-    case "Deal Fell Through":
-      style = "bg-red-100 text-red-600";
-      break;
-    default:
-      style = "bg-gray-100 text-gray-500";
-  }
-
-  const sizeClass = size === "xs" ? "text-xs" : "text-sm";
-
-  return (
-    <span className={`${sizeClass} px-2 py-1 rounded-md font-medium ${style}`}>
-      {label}
-    </span>
-  );
-};
-
-const formatAddress = (address: ListingResult["address"]) => {
-  if (!address) return "Address not available";
-
-  const parts = [
-    address.streetNumber,
-    address.streetDirection,
-    address.streetName,
-    address.streetSuffix,
-  ].filter(Boolean);
-
-  const street = parts.join(" ");
-  const cityState = [address.city, address.state].filter(Boolean).join(", ");
-
-  return [street, cityState].filter(Boolean).join(", ");
-};
-
-// ListingPreview component based on tooltip variant
-function ListingPreview({ listing, onClick }: ListingPreviewProps) {
-  return (
-    <div
-      className="p-4 cursor-pointer hover:bg-gray-50 transition-colors duration-200"
-      onClick={onClick}
-    >
-      <div className="flex items-start gap-4">
-        {/* Property Image */}
-        <div className="bg-gray-100 rounded-md overflow-hidden flex-shrink-0 h-20 aspect-[4/3]">
-          {listing.images?.[0] ? (
-            <img
-              src={`https://cdn.repliers.io/${listing.images[0]}?class=small`}
-              alt="Property"
-              className="object-cover w-full h-full"
-            />
-          ) : (
-            <div className="flex items-center justify-center h-full text-gray-400 text-sm">
-              No Image
-            </div>
-          )}
-        </div>
-
-        {/* Property Info */}
-        <div className="flex-grow min-w-0">
-          {/* Price & MLS */}
-          <div className="flex justify-between items-start mb-2">
-            <div className="flex items-center gap-2 min-w-0">
-              <h4 className="font-semibold text-gray-800 truncate text-base">
-                {listing.listPrice
-                  ? formatPrice(listing.listPrice)
-                  : "Price N/A"}
-              </h4>
-              <span className="text-gray-400 text-sm">|</span>
-              <span className="text-gray-600 truncate text-sm">
-                {listing.mlsNumber}
-              </span>
-            </div>
-            <div className="flex-shrink-0 ml-2">
-              <StatusTag
-                type={listing.type}
-                lastStatus={listing.lastStatus}
-                size="sm"
-              />
-            </div>
-          </div>
-
-          {/* Address */}
-          <p className="text-gray-600 leading-relaxed text-sm mb-2">
-            {formatAddress(listing.address)}
-          </p>
-
-          {/* Property Details */}
-          <div className="flex items-center text-gray-500 space-x-3 flex-wrap text-sm">
-            <span className="flex items-center gap-1">
-              <BedDouble className="w-4 h-4" />
-              {formatBedrooms(listing.details)}
-            </span>
-            <span className="flex items-center gap-1">
-              <Bath className="w-4 h-4" />
-              {formatBathrooms(listing.details)}
-            </span>
-            <span className="flex items-center gap-1">
-              <Car className="w-4 h-4" />
-              {formatParking(listing.details)}
-            </span>
-            {listing.details?.propertyType && (
-              <span className="text-gray-400">| {listing.details.propertyType}</span>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-interface PropertyTooltipProps {
-  open: boolean;
-  onClose: () => void;
-  listing: ListingResult | null;
-  position: {
-    x: number;
-    y: number;
-  };
-  mapContainer: HTMLElement | null;
-}
-
-interface ClusterTooltipProps {
-  open: boolean;
-  onClose: () => void;
-  properties: ListingResult[];
-  position: {
-    x: number;
-    y: number;
-  };
-  mapContainer: HTMLElement | null;
-}
-
-
-function PropertyTooltip({
-  open,
-  onClose,
-  listing,
-  position,
-  mapContainer,
-}: PropertyTooltipProps) {
-  const tooltipRef = useRef<HTMLDivElement>(null);
-  const [tooltipPosition, setTooltipPosition] = useState({ x: 0, y: 0 });
-  const [arrowPosition, setArrowPosition] = useState<'top' | 'bottom' | 'left' | 'right'>('bottom');
-  const [isPositioned, setIsPositioned] = useState(false);
-
-  // Reset positioned state when opening/closing
-  useEffect(() => {
-    setIsPositioned(false);
-  }, [open]);
-
-  // Calculate optimal tooltip position
-  useEffect(() => {
-    if (!open || !tooltipRef.current || !mapContainer) return;
-
-    const tooltip = tooltipRef.current;
-    const tooltipRect = tooltip.getBoundingClientRect();
-
-    // Convert marker position to viewport coordinates
-    const markerX = position.x;
-    const markerY = position.y;
-
-    // Tooltip dimensions - optimized for content
-    const tooltipWidth = 450; // Optimized width for better content fit
-    const tooltipHeight = tooltipRect.height || 200; // Estimated height
-
-    // Viewport boundaries (with padding)
-    const padding = 16;
-    const viewportWidth = window.innerWidth;
-    const viewportHeight = window.innerHeight;
-
-    // Calculate preferred position (above the marker)
-    let x = markerX - tooltipWidth / 2;
-    let y = markerY - tooltipHeight - 16; // 16px gap for arrow
-    let arrow: 'top' | 'bottom' | 'left' | 'right' = 'bottom';
-
-    // Horizontal boundary checks
-    if (x < padding) {
-      x = padding;
-    } else if (x + tooltipWidth > viewportWidth - padding) {
-      x = viewportWidth - tooltipWidth - padding;
-    }
-
-    // Vertical boundary checks
-    if (y < padding) {
-      // Not enough space above, show below marker
-      y = markerY + 32; // 32px gap for arrow + marker height
-      arrow = 'top';
-    }
-
-    // Final boundary check for bottom positioning
-    if (y + tooltipHeight > viewportHeight - padding) {
-      // Try positioning to the side
-      if (markerX > viewportWidth / 2) {
-        // Show on left side
-        x = markerX - tooltipWidth - 16;
-        y = markerY - tooltipHeight / 2;
-        arrow = 'right';
-      } else {
-        // Show on right side
-        x = markerX + 16;
-        y = markerY - tooltipHeight / 2;
-        arrow = 'left';
-      }
-
-      // Vertical centering adjustments for side positioning
-      if (y < padding) y = padding;
-      if (y + tooltipHeight > viewportHeight - padding) {
-        y = viewportHeight - tooltipHeight - padding;
-      }
-    }
-
-    setTooltipPosition({ x, y });
-    setArrowPosition(arrow);
-    setIsPositioned(true);
-  }, [open, position, mapContainer]);
-
-  // Handle click outside to close
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (tooltipRef.current && !tooltipRef.current.contains(event.target as Node)) {
-        onClose();
-      }
-    };
-
-    if (open) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [open, onClose]);
-
-  if (!open || !listing) return null;
-
-  // Arrow styles based on position
-  const getArrowStyles = () => {
-    const baseClasses = "absolute w-3 h-3 bg-white transform rotate-45 border";
-
-    switch (arrowPosition) {
-      case 'bottom':
-        return `${baseClasses} -bottom-1.5 left-1/2 -translate-x-1/2 border-r border-b border-gray-200`;
-      case 'top':
-        return `${baseClasses} -top-1.5 left-1/2 -translate-x-1/2 border-l border-t border-gray-200`;
-      case 'left':
-        return `${baseClasses} -left-1.5 top-1/2 -translate-y-1/2 border-l border-b border-gray-200`;
-      case 'right':
-        return `${baseClasses} -right-1.5 top-1/2 -translate-y-1/2 border-r border-t border-gray-200`;
-      default:
-        return `${baseClasses} -bottom-1.5 left-1/2 -translate-x-1/2 border-r border-b border-gray-200`;
-    }
-  };
-
-  return (
-    <div
-      ref={tooltipRef}
-      className="fixed w-[450px]"
-      style={{
-        zIndex: 9999,
-        left: `${tooltipPosition.x}px`,
-        top: `${tooltipPosition.y}px`,
-        visibility: isPositioned ? 'visible' : 'hidden',
-        opacity: isPositioned ? 1 : 0,
-        transition: 'opacity 150ms ease-in-out',
-      }}
-    >
-      {/* Tooltip Card */}
-      <div className="bg-white rounded-lg shadow-lg border border-gray-200 overflow-hidden relative">
-        {/* Arrow pointer */}
-        <div className={getArrowStyles()} />
-
-        {/* Close button */}
-        <button
-          onClick={onClose}
-          className="absolute top-2 right-2 z-10 p-1 rounded-full bg-white hover:bg-gray-100 transition-colors shadow-sm border border-gray-200"
-          aria-label="Close"
-        >
-          <X size={16} className="text-gray-600" />
-        </button>
-
-        {/* Content - Use ListingPreview component */}
-        <ListingPreview
-          listing={listing}
-          onClick={() => {
-            // TODO: Replace with actual navigation to full property details page
-            // Example: window.open(`/listings/${listing.mlsNumber}`, '_blank');
-            // Example: router.push(`/property/${listing.mlsNumber}`);
-          }}
-        />
-      </div>
-    </div>
-  );
-}
-
-function ClusterTooltip({
-  open,
-  onClose,
-  properties,
-  position,
-  mapContainer,
-}: ClusterTooltipProps) {
-  const tooltipRef = useRef<HTMLDivElement>(null);
-  const [tooltipPosition, setTooltipPosition] = useState({ x: 0, y: 0 });
-  const [arrowPosition, setArrowPosition] = useState<'top' | 'bottom' | 'left' | 'right'>('bottom');
-  const [isPositioned, setIsPositioned] = useState(false);
-
-  // Reset positioned state when opening/closing
-  useEffect(() => {
-    if (!open) {
-      setIsPositioned(false);
-      return;
-    }
-  }, [open]);
-
-  // Calculate optimal tooltip position immediately when opened
-  useEffect(() => {
-    if (!open || !mapContainer) {
-      return;
-    }
-
-    // Calculate position immediately with estimated dimensions
-    const markerX = position.x;
-    const markerY = position.y;
-
-    // Tooltip dimensions - use estimates to avoid getBoundingClientRect delay
-    const tooltipWidth = 480;
-    const tooltipHeight = Math.min(properties.length * 120 + 80, 400); // Estimate based on property count
-
-    // Viewport boundaries (with padding)
-    const padding = 16;
-    const viewportWidth = window.innerWidth;
-    const viewportHeight = window.innerHeight;
-
-    // Calculate preferred position (above the marker)
-    let x = markerX - tooltipWidth / 2;
-    let y = markerY - tooltipHeight - 16; // 16px gap for arrow
-    let arrow: 'top' | 'bottom' | 'left' | 'right' = 'bottom';
-
-    // Horizontal boundary checks
-    if (x < padding) {
-      x = padding;
-    } else if (x + tooltipWidth > viewportWidth - padding) {
-      x = viewportWidth - tooltipWidth - padding;
-    }
-
-    // Vertical boundary checks
-    if (y < padding) {
-      // Not enough space above, show below marker
-      y = markerY + 32; // 32px gap for arrow + marker height
-      arrow = 'top';
-    }
-
-    // Final boundary check for bottom positioning
-    if (y + tooltipHeight > viewportHeight - padding) {
-      // Try positioning to the side
-      if (markerX > viewportWidth / 2) {
-        // Show on left side
-        x = markerX - tooltipWidth - 16;
-        y = markerY - tooltipHeight / 2;
-        arrow = 'right';
-      } else {
-        // Show on right side
-        x = markerX + 16;
-        y = markerY - tooltipHeight / 2;
-        arrow = 'left';
-      }
-
-      // Vertical centering adjustments for side positioning
-      if (y < padding) y = padding;
-      if (y + tooltipHeight > viewportHeight - padding) {
-        y = viewportHeight - tooltipHeight - padding;
-      }
-    }
-
-    setTooltipPosition({ x, y });
-    setArrowPosition(arrow);
-    setIsPositioned(true);
-  }, [open, position, mapContainer, properties.length]);
-
-  // Handle click outside to close
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (tooltipRef.current && !tooltipRef.current.contains(event.target as Node)) {
-        onClose();
-      }
-    };
-
-    if (open) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [open, onClose]);
-
-  if (!open || properties.length === 0) return null;
-
-  // Arrow styles based on position
-  const getArrowStyles = () => {
-    const baseClasses = "absolute w-3 h-3 bg-white transform rotate-45 border";
-
-    switch (arrowPosition) {
-      case 'bottom':
-        return `${baseClasses} -bottom-1.5 left-1/2 -translate-x-1/2 border-r border-b border-gray-200`;
-      case 'top':
-        return `${baseClasses} -top-1.5 left-1/2 -translate-x-1/2 border-l border-t border-gray-200`;
-      case 'left':
-        return `${baseClasses} -left-1.5 top-1/2 -translate-y-1/2 border-l border-b border-gray-200`;
-      case 'right':
-        return `${baseClasses} -right-1.5 top-1/2 -translate-y-1/2 border-r border-t border-gray-200`;
-      default:
-        return `${baseClasses} -bottom-1.5 left-1/2 -translate-x-1/2 border-r border-b border-gray-200`;
-    }
-  };
-
-  return (
-    <div
-      ref={tooltipRef}
-      className="fixed w-[480px]"
-      style={{
-        zIndex: 9999,
-        left: `${tooltipPosition.x}px`,
-        top: `${tooltipPosition.y}px`,
-        visibility: isPositioned ? 'visible' : 'hidden',
-        opacity: isPositioned ? 1 : 0,
-        transition: 'opacity 150ms ease-in-out',
-      }}
-    >
-      {/* Tooltip Card */}
-      <div className="bg-white rounded-lg shadow-lg border border-gray-200 overflow-hidden relative">
-        {/* Arrow pointer */}
-        <div className={getArrowStyles()} />
-
-        {/* Header */}
-        <div className="px-4 py-3 border-b border-gray-100 bg-gray-50 relative">
-          <h3 className="text-sm font-semibold text-gray-800">
-            {properties.length === 1 ? '1 Property' : `${properties.length} Properties`}
-          </h3>
-
-          {/* Close button */}
-          <button
-            onClick={onClose}
-            className="absolute top-2 right-2 p-1 rounded-full bg-white hover:bg-gray-100 transition-colors shadow-sm border border-gray-200"
-            aria-label="Close"
-          >
-            <X size={16} className="text-gray-600" />
-          </button>
-        </div>
-
-        {/* Scrollable Content - Stack ListingPreview components */}
-        <div className="max-h-[400px] overflow-y-auto">
-          {properties.map((listing, index) => (
-            <div key={`${listing.mlsNumber}-${index}`} className="border-b border-gray-100 last:border-b-0">
-              <ListingPreview
-                listing={listing}
-                onClick={() => {
-                  // TODO: Replace with actual navigation to full property details page
-                  // Example: window.open(`/listings/${listing.mlsNumber}`, '_blank');
-                  // Example: router.push(`/property/${listing.mlsNumber}`);
-                }}
-              />
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Utility functions for auto-centering
-
-/**
- * Extract coordinates from various API response formats
- */
-const extractCoordinates = (listing: any): Coordinate | null => {
-  // Try different coordinate formats
-  if (listing.coordinates?.lat && listing.coordinates?.lng) {
-    return { lat: listing.coordinates.lat, lng: listing.coordinates.lng };
-  }
-  if (listing.latitude && listing.longitude) {
-    return { lat: listing.latitude, lng: listing.longitude };
-  }
-  if (listing.map?.latitude && listing.map?.longitude) {
-    return { lat: listing.map.latitude, lng: listing.map.longitude };
-  }
-  // Also check cluster coordinate formats
-  if (listing.location?.latitude && listing.location?.longitude) {
-    return { lat: listing.location.latitude, lng: listing.location.longitude };
-  }
-  if (listing.center?.lat && listing.center?.lng) {
-    return { lat: listing.center.lat, lng: listing.center.lng };
-  }
-  return null;
-};
-
-/**
- * Calculate bounds from an array of coordinates
- */
-const getPolygonBounds = (coordinates: Coordinate[]): MapBounds | null => {
-  if (coordinates.length === 0) return null;
-
-  let north = coordinates[0].lat;
-  let south = coordinates[0].lat;
-  let east = coordinates[0].lng;
-  let west = coordinates[0].lng;
-
-  coordinates.forEach((coord) => {
-    north = Math.max(north, coord.lat);
-    south = Math.min(south, coord.lat);
-    east = Math.max(east, coord.lng);
-    west = Math.min(west, coord.lng);
-  });
-
-  return { north, south, east, west };
-};
-
-/**
- * Calculate density-weighted center from coordinates (original grid method)
- */
-const calculateAverageCenter = (
-  coordinates: Coordinate[],
-  bounds: MapBounds
-): [number, number] => {
-  const { north, south, east, west } = bounds;
-
-  // Create a density grid to find the densest area
-  const gridSize = 20; // 20x20 grid for density calculation
-  const latStep = (north - south) / gridSize;
-  const lngStep = (east - west) / gridSize;
-
-  // Initialize density grid
-  const densityGrid: number[][] = Array(gridSize)
-    .fill(0)
-    .map(() => Array(gridSize).fill(0));
-
-  // Count listings in each grid cell
-  coordinates.forEach((coord) => {
-    const latIndex = Math.floor((coord.lat - south) / latStep);
-    const lngIndex = Math.floor((coord.lng - west) / lngStep);
-
-    // Ensure indices are within bounds
-    const safeLatIndex = Math.max(0, Math.min(gridSize - 1, latIndex));
-    const safeLngIndex = Math.max(0, Math.min(gridSize - 1, lngIndex));
-
-    densityGrid[safeLatIndex][safeLngIndex]++;
-  });
-
-  // Find the densest grid cell
-  let maxDensity = 0;
-  let densestLat = 0;
-  let densestLng = 0;
-
-  for (let i = 0; i < gridSize; i++) {
-    for (let j = 0; j < gridSize; j++) {
-      if (densityGrid[i][j] > maxDensity) {
-        maxDensity = densityGrid[i][j];
-        densestLat = south + (i + 0.5) * latStep; // Center of grid cell
-        densestLng = west + (j + 0.5) * lngStep;
-      }
-    }
-  }
-
-  // If no dense area found, fall back to geometric center
-  const center: [number, number] =
-    maxDensity > 0
-      ? [densestLng, densestLat]
-      : [(east + west) / 2, (north + south) / 2];
-
-  return center;
-};
-
-
-/**
- * Cache management for auto-center data
- */
-const getCachedAutoCenter = (apiKey: string): AutoCenterData | null => {
-  try {
-    const cached = localStorage.getItem(`mapListings-autoCenter-${apiKey}`);
-    if (cached) {
-      const data = JSON.parse(cached);
-      // Check if cache is less than 24 hours old
-      if (Date.now() - data.timestamp < 24 * 60 * 60 * 1000) {
-        return { center: data.center };
-      }
-    }
-  } catch (error) {
-  }
-  return null;
-};
-
-const setCachedAutoCenter = (apiKey: string, data: AutoCenterData): void => {
-  try {
-    const cacheData = {
-      ...data,
-      timestamp: Date.now(),
-    };
-    localStorage.setItem(
-      `mapListings-autoCenter-${apiKey}`,
-      JSON.stringify(cacheData)
-    );
-  } catch (error) {
-  }
-};
-
-/**
- * Auto-detect optimal map center from listings data using bundled searches
- */
-const detectAutoCenter = async (
-  apiKey: string,
-  method: "average" | "city" = "average"
-): Promise<AutoCenterData | null> => {
-  // Try bundled approach first, fallback to individual calls if 400 error
-  try {
-
-    // Use bundled search to get both clusters and listings in one request
-    const bundledQuery = {
-      queries: [
-        {
-          cluster: true,
-          clusterPrecision: 8, // Low precision for bigger clusters
-          clusterLimit: 50,
-          status: "A"
-        },
-        {
-          cluster: false,
-          listings: true,
-          pageSize: 500, // Get more listings for better center calculation
-          status: "A"
-        }
-      ]
-    };
-
-    const response = await fetch("https://api.repliers.io/listings", {
-      method: "POST",
-      headers: {
-        "REPLIERS-API-KEY": apiKey,
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(bundledQuery),
-    });
-
-    if (response.status === 400) {
-      return detectAutoCenterFallback(apiKey, method);
-    }
-
-    if (!response.ok) {
-      throw new Error(
-        `Bundled auto-center API Error: ${response.status} ${response.statusText}`
-      );
-    }
-
-    const data = await response.json();
-
-    // Extract clusters and listings from bundled response
-    const clusterData = data[0] || {}; // First query result (clusters)
-    const listingData = data[1] || {}; // Second query result (listings)
-
-    const clusters = clusterData.aggregates?.map?.clusters || clusterData.clusters || [];
-    const listings = listingData.listings || listingData.results || [];
-
-    if (method === "city" && clusters.length > 0) {
-      // Find the cluster with the most listings
-      let largestCluster = null;
-      let maxCount = 0;
-
-      clusters.forEach((cluster: any) => {
-        const count = cluster.count || 0;
-        if (count > maxCount) {
-          maxCount = count;
-          largestCluster = cluster;
-        }
-      });
-
-      if (largestCluster) {
-        // Extract coordinates from the largest cluster
-        const coordinates = [
-          (largestCluster as any).location?.longitude ||
-            (largestCluster as any).coordinates?.lng ||
-            (largestCluster as any).longitude ||
-            (largestCluster as any).center?.lng,
-          (largestCluster as any).location?.latitude ||
-            (largestCluster as any).coordinates?.lat ||
-            (largestCluster as any).latitude ||
-            (largestCluster as any).center?.lat,
-        ];
-
-        if (coordinates[0] && coordinates[1]) {
-          const center: [number, number] = [coordinates[0], coordinates[1]];
-          const autoCenter: AutoCenterData = { center };
-
-          // Cache the result with method-specific key
-          setCachedAutoCenter(`${apiKey}-${method}`, autoCenter);
-          return autoCenter;
-        }
-      }
-
-    }
-
-    // Average method (or fallback)
-    if (listings.length === 0) {
-      return null;
-    }
-
-    // Extract coordinates from listings
-    const coordinates: Coordinate[] = [];
-    listings.forEach((listing: any) => {
-      const coord = extractCoordinates(listing);
-      if (coord) {
-        coordinates.push(coord);
-      }
-    });
-
-    if (coordinates.length === 0) {
-      return null;
-    }
-
-    // Calculate bounds and center using average method
-    const bounds = getPolygonBounds(coordinates);
-    if (!bounds) {
-      return null;
-    }
-
-    const center = calculateAverageCenter(coordinates, bounds);
-    const autoCenter: AutoCenterData = { center };
-
-    // Cache the result with method-specific key
-    setCachedAutoCenter(`${apiKey}-${method}`, autoCenter);
-
-    return autoCenter;
-  } catch (error) {
-    // Try fallback approach for non-400 errors too
-    return detectAutoCenterFallback(apiKey, method);
-  }
-};
-
-/**
- * Fallback auto-center detection using individual API calls
- */
-const detectAutoCenterFallback = async (
-  apiKey: string,
-  method: "average" | "city" = "average"
-): Promise<AutoCenterData | null> => {
-  try {
-    let clusters: any[] = [];
-    let listings: any[] = [];
-
-    // First, try to get clusters for city method
-    if (method === "city") {
-      try {
-        const clusterUrl = new URL("https://api.repliers.io/listings");
-        clusterUrl.searchParams.set("cluster", "true");
-        clusterUrl.searchParams.set("clusterPrecision", "8");
-        clusterUrl.searchParams.set("clusterLimit", "50");
-        clusterUrl.searchParams.set("status", "A");
-
-        const clusterResponse = await fetch(clusterUrl.toString(), {
-          headers: {
-            "REPLIERS-API-KEY": apiKey,
-            "Content-Type": "application/json",
-          },
-        });
-
-        if (clusterResponse.ok) {
-          const clusterData = await clusterResponse.json();
-          clusters = clusterData.aggregates?.map?.clusters || clusterData.clusters || [];
-        }
-      } catch (error) {
-        // Cluster fetch failed in fallback
-      }
-    }
-
-    // Get individual listings for average method or as fallback for city method
-    try {
-      const listingUrl = new URL("https://api.repliers.io/listings");
-      listingUrl.searchParams.set("pageSize", "500");
-      listingUrl.searchParams.set("status", "A");
-
-      const listingResponse = await fetch(listingUrl.toString(), {
-        headers: {
-          "REPLIERS-API-KEY": apiKey,
-          "Content-Type": "application/json",
-        },
-      });
-
-      if (listingResponse.ok) {
-        const listingData = await listingResponse.json();
-        listings = listingData.listings || listingData.results || [];
-      }
-    } catch (error) {
-      // Listing fetch failed in fallback
-    }
-
-    // Process the same way as bundled approach
-    if (method === "city" && clusters.length > 0) {
-      // Find the cluster with the most listings
-      let largestCluster = null;
-      let maxCount = 0;
-
-      clusters.forEach((cluster: any) => {
-        const count = cluster.count || 0;
-        if (count > maxCount) {
-          maxCount = count;
-          largestCluster = cluster;
-        }
-      });
-
-      if (largestCluster) {
-        // Extract coordinates from the largest cluster
-        const coordinates = [
-          (largestCluster as any).location?.longitude ||
-            (largestCluster as any).coordinates?.lng ||
-            (largestCluster as any).longitude ||
-            (largestCluster as any).center?.lng,
-          (largestCluster as any).location?.latitude ||
-            (largestCluster as any).coordinates?.lat ||
-            (largestCluster as any).latitude ||
-            (largestCluster as any).center?.lat,
-        ];
-
-        if (coordinates[0] && coordinates[1]) {
-          const center: [number, number] = [coordinates[0], coordinates[1]];
-          const autoCenter: AutoCenterData = { center };
-
-          // Cache the result with method-specific key
-          setCachedAutoCenter(`${apiKey}-${method}`, autoCenter);
-          return autoCenter;
-        }
-      }
-
-    }
-
-    // Average method (or fallback)
-    if (listings.length === 0) {
-      return null;
-    }
-
-    // Extract coordinates from listings
-    const coordinates: Coordinate[] = [];
-    listings.forEach((listing: any) => {
-      const coord = extractCoordinates(listing);
-      if (coord) {
-        coordinates.push(coord);
-      }
-    });
-
-    if (coordinates.length === 0) {
-      return null;
-    }
-
-    // Calculate bounds and center using average method
-    const bounds = getPolygonBounds(coordinates);
-    if (!bounds) {
-      return null;
-    }
-
-    const center = calculateAverageCenter(coordinates, bounds);
-    const autoCenter: AutoCenterData = { center };
-
-    // Cache the result with method-specific key
-    setCachedAutoCenter(`${apiKey}-${method}`, autoCenter);
-
-    return autoCenter;
-  } catch (error) {
-    return null;
-  }
-};
-
-
-// Property Type Filter Component
-interface PropertyTypeFilterProps {
-  filters: MapFilters;
-  onFiltersChange: (filters: MapFilters) => void;
-  apiKey: string;
-}
-
-interface PriceRangeFilterProps {
-  isOpen: boolean;
-  initialMin: number;
-  initialMax: number | null; // null represents "Max" (no limit)
-  onApply: (min: number, max: number | null) => void;
-  onCancel: () => void;
-  priceBreakpoints?: number[]; // Optional custom breakpoints
-}
-
-function PropertyTypeFilter({ filters, onFiltersChange, apiKey }: PropertyTypeFilterProps) {
-  const [propertyTypes, setPropertyTypes] = useState<string[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-
-  // Fetch available property types on mount
-  useEffect(() => {
-    const fetchPropertyTypes = async () => {
-      try {
-        const url = new URL("https://api.repliers.io/listings");
-        url.searchParams.set("aggregates", "details.propertyType");
-        url.searchParams.set("listings", "false");
-        url.searchParams.set("status", "A");
-
-        const response = await fetch(url.toString(), {
-          headers: {
-            "REPLIERS-API-KEY": apiKey,
-            "Content-Type": "application/json",
-          },
-        });
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch property types: ${response.status}`);
-        }
-
-        const data = await response.json();
-        const propertyTypeData = data.aggregates?.details?.propertyType || {};
-
-        // Extract property type names and sort by count (descending)
-        const sortedPropertyTypes = Object.entries(propertyTypeData)
-          .sort(([, a], [, b]) => (b as number) - (a as number))
-          .map(([propertyType]) => propertyType);
-
-        setPropertyTypes(sortedPropertyTypes);
-      } catch (error) {
-        // Silently handle property type fetch errors
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchPropertyTypes();
-  }, [apiKey]);
-
-  const handlePropertyTypeToggle = (propertyType: string) => {
-    const isSelected = filters.propertyTypes.includes(propertyType);
-    const newPropertyTypes = isSelected
-      ? filters.propertyTypes.filter(type => type !== propertyType)
-      : [...filters.propertyTypes, propertyType];
-
-    onFiltersChange({
-      ...filters,
-      propertyTypes: newPropertyTypes,
-    });
-  };
-
-  const handleAllToggle = () => {
-    const allSelected = filters.propertyTypes.length === propertyTypes.length;
-    onFiltersChange({
-      ...filters,
-      propertyTypes: allSelected ? [] : [...propertyTypes],
-    });
-  };
-
-  if (loading) {
-    return (
-      <div style={{ position: "relative", marginBottom: "12px" }}>
-        <div style={{
-          padding: "8px 12px",
-          fontSize: "12px",
-          color: "#6b7280",
-          fontStyle: "italic"
-        }}>
-          Loading property types...
-        </div>
-      </div>
-    );
-  }
-
-  const allSelected = filters.propertyTypes.length === 0 || filters.propertyTypes.length === propertyTypes.length;
-  const someSelected = filters.propertyTypes.length > 0 && filters.propertyTypes.length < propertyTypes.length;
-
-  return (
-    <div style={{ position: "relative", marginBottom: "12px" }}>
-      <button
-        onClick={() => setIsDropdownOpen(!isDropdownOpen)}
-        style={{
-          width: "100%",
-          padding: "8px 12px",
-          backgroundColor: "white",
-          border: "1px solid #d1d5db",
-          borderRadius: "6px",
-          fontSize: "14px",
-          cursor: "pointer",
-          display: "flex",
-          justifyContent: "space-between",
-          alignItems: "center",
-          textAlign: "left",
-        }}
-      >
-        <span>
-          {allSelected
-            ? "All property types"
-            : someSelected
-            ? `${filters.propertyTypes.length} selected`
-            : "Select property types"}
-        </span>
-        <ChevronDown
-          size={16}
-          style={{
-            transform: isDropdownOpen ? "rotate(180deg)" : "rotate(0deg)",
-            transition: "transform 0.2s"
-          }}
-        />
-      </button>
-
-      {isDropdownOpen && (
-        <div
-          style={{
-            position: "absolute",
-            top: "100%",
-            left: 0,
-            right: 0,
-            backgroundColor: "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            marginTop: "4px",
-            boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-            zIndex: 1002,
-            maxHeight: "300px",
-            overflowY: "auto",
-          }}
-        >
-          {/* All property types option */}
-          <div
-            onClick={handleAllToggle}
-            style={{
-              padding: "8px 12px",
-              fontSize: "14px",
-              cursor: "pointer",
-              display: "flex",
-              alignItems: "center",
-              gap: "8px",
-              backgroundColor: allSelected ? "#f3f4f6" : "white",
-              borderBottom: "1px solid #f3f4f6",
-            }}
-            onMouseEnter={(e) => {
-              if (!allSelected) {
-                e.currentTarget.style.backgroundColor = "#f9fafb";
-              }
-            }}
-            onMouseLeave={(e) => {
-              if (!allSelected) {
-                e.currentTarget.style.backgroundColor = "white";
-              }
-            }}
-          >
-            <div style={{
-              width: "16px",
-              height: "16px",
-              border: "2px solid #d1d5db",
-              borderRadius: "3px",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              backgroundColor: allSelected ? "#10b981" : "white",
-              borderColor: allSelected ? "#10b981" : "#d1d5db",
-            }}>
-              {allSelected && (
-                <svg
-                  width="10"
-                  height="8"
-                  viewBox="0 0 10 8"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M8.5 1L3.5 6L1.5 4"
-                    stroke="white"
-                    strokeWidth="1.5"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  />
-                </svg>
-              )}
-            </div>
-            All property types
-          </div>
-
-          {/* Individual property type options */}
-          {propertyTypes.map((propertyType) => {
-            const isSelected = filters.propertyTypes.includes(propertyType);
-            return (
-              <div
-                key={propertyType}
-                onClick={() => handlePropertyTypeToggle(propertyType)}
-                style={{
-                  padding: "8px 12px",
-                  fontSize: "14px",
-                  cursor: "pointer",
-                  display: "flex",
-                  alignItems: "center",
-                  gap: "8px",
-                  backgroundColor: isSelected ? "#f3f4f6" : "white",
-                }}
-                onMouseEnter={(e) => {
-                  if (!isSelected) {
-                    e.currentTarget.style.backgroundColor = "#f9fafb";
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (!isSelected) {
-                    e.currentTarget.style.backgroundColor = "white";
-                  }
-                }}
-              >
-                <div style={{
-                  width: "16px",
-                  height: "16px",
-                  border: "2px solid #d1d5db",
-                  borderRadius: "3px",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  backgroundColor: isSelected ? "#10b981" : "white",
-                  borderColor: isSelected ? "#10b981" : "#d1d5db",
-                }}>
-                  {isSelected && (
-                    <svg
-                      width="10"
-                      height="8"
-                      viewBox="0 0 10 8"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M8.5 1L3.5 6L1.5 4"
-                        stroke="white"
-                        strokeWidth="1.5"
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                      />
-                    </svg>
-                  )}
-                </div>
-                {propertyType}
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Price Range Filter Component
-function PriceRangeFilter({
-  isOpen,
-  initialMin,
-  initialMax,
-  onApply,
-  onCancel,
-  priceBreakpoints = [0, 500000, 1000000, 2000000, 4000000, Infinity]
-}: PriceRangeFilterProps) {
-  const [tempMinPrice, setTempMinPrice] = useState(initialMin);
-  const [tempMaxPrice, setTempMaxPrice] = useState(initialMax);
-  const [isDragging, setIsDragging] = useState<'min' | 'max' | null>(null);
-  const sliderRef = useRef<HTMLDivElement>(null);
-
-  // Reset temp values when modal opens/closes or initial values change
-  useEffect(() => {
-    setTempMinPrice(initialMin);
-    setTempMaxPrice(initialMax);
-  }, [isOpen, initialMin, initialMax]);
-
-  // Format price for display
-  const formatDisplayPrice = useCallback((price: number | null): string => {
-    if (price === null || price === Infinity) return "Max";
-    if (price === 0) return "$0";
-    if (price >= 1000000) {
-      const millions = price / 1000000;
-      return millions >= 10 ? `$${Math.round(millions)}M` : `$${Math.round(millions * 10) / 10}M`;
-    }
-    if (price >= 1000) {
-      const thousands = price / 1000;
-      return `$${Math.round(thousands)}K`;
-    }
-    return `$${Math.round(price)}`;
-  }, []);
-
-  // Get slider position as percentage
-  const getSliderPosition = useCallback((price: number | null): number => {
-    if (price === null || price === Infinity) return 100;
-
-    const maxFinitePrice = priceBreakpoints[priceBreakpoints.length - 2];
-    if (price >= maxFinitePrice) return 100;
-    if (price <= 0) return 0;
-
-    // Find position between breakpoints
-    for (let i = 0; i < priceBreakpoints.length - 1; i++) {
-      const current = priceBreakpoints[i];
-      const next = priceBreakpoints[i + 1];
-
-      if (price >= current && price <= next) {
-        const segmentSize = 100 / (priceBreakpoints.length - 1);
-        const segmentStart = i * segmentSize;
-        const progressInSegment = next === Infinity ? 0 : (price - current) / (next - current);
-        return segmentStart + (progressInSegment * segmentSize);
-      }
-    }
-
-    return 0;
-  }, [priceBreakpoints]);
-
-  // Get price from slider position
-  const getPriceFromPosition = useCallback((position: number): number | null => {
-    if (position >= 100) return null; // Max
-    if (position <= 0) return 0;
-
-    const segmentSize = 100 / (priceBreakpoints.length - 1);
-    const segmentIndex = Math.floor(position / segmentSize);
-    const progressInSegment = (position % segmentSize) / segmentSize;
-
-    if (segmentIndex >= priceBreakpoints.length - 2) return null;
-
-    const current = priceBreakpoints[segmentIndex];
-    const next = priceBreakpoints[segmentIndex + 1];
-
-    if (next === Infinity) return null;
-
-    return Math.round(current + (progressInSegment * (next - current)));
-  }, [priceBreakpoints]);
-
-  // Handle mouse events for dragging
-  const handleMouseDown = useCallback((handle: 'min' | 'max') => (e: React.MouseEvent) => {
-    e.preventDefault();
-    setIsDragging(handle);
-  }, []);
-
-  const handleMouseMove = useCallback((e: MouseEvent) => {
-    if (!isDragging || !sliderRef.current) return;
-
-    const rect = sliderRef.current.getBoundingClientRect();
-    const position = Math.max(0, Math.min(100, ((e.clientX - rect.left) / rect.width) * 100));
-    const newPrice = getPriceFromPosition(position);
-
-    if (isDragging === 'min') {
-      const maxValue = tempMaxPrice === null ? Infinity : tempMaxPrice;
-      const constrainedPrice = newPrice === null ? maxValue : Math.min(newPrice, maxValue);
-      setTempMinPrice(constrainedPrice === Infinity ? 0 : constrainedPrice);
-    } else if (isDragging === 'max') {
-      const constrainedPrice = newPrice === null ? null : Math.max(newPrice, tempMinPrice);
-      setTempMaxPrice(constrainedPrice);
-    }
-  }, [isDragging, tempMinPrice, tempMaxPrice, getPriceFromPosition]);
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(null);
-  }, []);
-
-  // Add/remove global mouse event listeners
-  useEffect(() => {
-    if (isDragging) {
-      document.addEventListener('mousemove', handleMouseMove);
-      document.addEventListener('mouseup', handleMouseUp);
-      return () => {
-        document.removeEventListener('mousemove', handleMouseMove);
-        document.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [isDragging, handleMouseMove, handleMouseUp]);
-
-  // Handle keyboard events for accessibility
-  const handleKeyDown = useCallback((handle: 'min' | 'max') => (e: React.KeyboardEvent) => {
-    if (e.key === 'ArrowLeft' || e.key === 'ArrowRight') {
-      e.preventDefault();
-      const direction = e.key === 'ArrowLeft' ? -1 : 1;
-      const currentPrice = handle === 'min' ? tempMinPrice : tempMaxPrice;
-      const currentPos = getSliderPosition(currentPrice);
-      const newPos = Math.max(0, Math.min(100, currentPos + (direction * 2)));
-      const newPrice = getPriceFromPosition(newPos);
-
-      if (handle === 'min') {
-        const maxValue = tempMaxPrice === null ? Infinity : tempMaxPrice;
-        const constrainedPrice = newPrice === null ? maxValue : Math.min(newPrice, maxValue);
-        setTempMinPrice(constrainedPrice === Infinity ? 0 : constrainedPrice);
-      } else {
-        const constrainedPrice = newPrice === null ? null : Math.max(newPrice || 0, tempMinPrice);
-        setTempMaxPrice(constrainedPrice);
-      }
-    }
-  }, [tempMinPrice, tempMaxPrice, getSliderPosition, getPriceFromPosition]);
-
-  // Handle ESC key to close modal
-  useEffect(() => {
-    const handleEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' && isOpen) {
-        onCancel();
-      }
-    };
-
-    if (isOpen) {
-      document.addEventListener('keydown', handleEsc);
-      return () => document.removeEventListener('keydown', handleEsc);
-    }
-  }, [isOpen, onCancel]);
-
-  if (!isOpen) return null;
-
-  const minPosition = getSliderPosition(tempMinPrice);
-  const maxPosition = getSliderPosition(tempMaxPrice);
-
-  return (
-    <div
-      style={{
-        position: "absolute",
-        top: "100%",
-        left: 0,
-        width: "400px", // Fixed width instead of constrained by parent
-        backgroundColor: "white",
-        border: "1px solid #d1d5db",
-        borderRadius: "6px",
-        marginTop: "4px",
-        boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-        zIndex: 1002,
-        padding: "20px",
-      }}
-    >
-      {/* Current Selection Display */}
-      <div style={{
-        textAlign: 'center',
-        fontSize: '16px',
-        fontWeight: '600',
-        color: '#374151',
-        marginBottom: '20px',
-      }}>
-        {formatDisplayPrice(tempMinPrice)} - {formatDisplayPrice(tempMaxPrice)}
-      </div>
-
-      {/* Slider Container */}
-      <div style={{ marginBottom: '20px' }}>
-        <div
-          ref={sliderRef}
-          style={{
-            position: 'relative',
-            height: '8px',
-            backgroundColor: '#e5e7eb',
-            borderRadius: '4px',
-            marginBottom: '16px',
-          }}
-        >
-          {/* Active Range */}
-          <div
-            style={{
-              position: 'absolute',
-              top: 0,
-              left: `${minPosition}%`,
-              width: `${maxPosition - minPosition}%`,
-              height: '100%',
-              backgroundColor: '#3b82f6',
-              borderRadius: '4px',
-            }}
-          />
-
-          {/* Min Handle */}
-          <div
-            style={{
-              position: 'absolute',
-              left: `${minPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '20px',
-              height: '20px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'min' ? 'grabbing' : 'grab',
-              boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('min')}
-            onKeyDown={handleKeyDown('min')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Minimum price"
-            aria-valuemin={0}
-            aria-valuemax={tempMaxPrice === null ? priceBreakpoints[priceBreakpoints.length - 2] : tempMaxPrice}
-            aria-valuenow={tempMinPrice}
-            aria-valuetext={formatDisplayPrice(tempMinPrice)}
-          />
-
-          {/* Max Handle */}
-          <div
-            style={{
-              position: 'absolute',
-              left: `${maxPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '20px',
-              height: '20px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'max' ? 'grabbing' : 'grab',
-              boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('max')}
-            onKeyDown={handleKeyDown('max')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Maximum price"
-            aria-valuemin={tempMinPrice}
-            aria-valuemax={priceBreakpoints[priceBreakpoints.length - 2]}
-            aria-valuenow={tempMaxPrice === null ? priceBreakpoints[priceBreakpoints.length - 2] : tempMaxPrice}
-            aria-valuetext={formatDisplayPrice(tempMaxPrice)}
-          />
-        </div>
-
-        {/* Price Markers */}
-        <div style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          fontSize: '12px',
-          color: '#6b7280',
-        }}>
-          {priceBreakpoints.slice(0, -1).map((price, index) => (
-            <span key={index}>
-              {formatDisplayPrice(price)}
-            </span>
-          ))}
-          <span>Max</span>
-        </div>
-      </div>
-
-      {/* Action Buttons */}
-      <div style={{
-        display: 'flex',
-        gap: '12px',
-        justifyContent: 'flex-end',
-      }}>
-        <button
-          onClick={onCancel}
-          style={{
-            padding: '8px 16px',
-            fontSize: '14px',
-            fontWeight: '500',
-            color: '#374151',
-            backgroundColor: 'white',
-            border: '1px solid #d1d5db',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Cancel
-        </button>
-        <button
-          onClick={() => onApply(tempMinPrice, tempMaxPrice)}
-          style={{
-            padding: '8px 16px',
-            fontSize: '14px',
-            fontWeight: '500',
-            color: 'white',
-            backgroundColor: '#3b82f6',
-            border: 'none',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Apply
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Square Footage Filter Component
-interface SquareFootageFilterProps {
-  isOpen: boolean;
-  initialMin: number;
-  initialMax: number | null; // null represents "Max" (no limit)
-  onApply: (min: number, max: number | null) => void;
-  onCancel: () => void;
-  sqftBreakpoints?: number[]; // Optional custom breakpoints
-}
-
-function SquareFootageFilter({
-  isOpen,
-  initialMin,
-  initialMax,
-  onApply,
-  onCancel,
-  sqftBreakpoints = [0, 1000, 2000, 3000, 5000, Infinity]
-}: SquareFootageFilterProps) {
-  const [tempMinSqft, setTempMinSqft] = useState(initialMin);
-  const [tempMaxSqft, setTempMaxSqft] = useState(initialMax);
-  const [isDragging, setIsDragging] = useState<'min' | 'max' | null>(null);
-  const sliderRef = useRef<HTMLDivElement>(null);
-
-  // Reset temp values when modal opens/closes or initial values change
-  useEffect(() => {
-    setTempMinSqft(initialMin);
-    setTempMaxSqft(initialMax);
-  }, [isOpen, initialMin, initialMax]);
-
-  // Format sqft for display
-  const formatDisplaySqft = useCallback((sqft: number | null): string => {
-    if (sqft === null || sqft === Infinity) return "Max";
-    if (sqft === 0) return "0";
-    if (sqft >= 1000) {
-      const thousands = sqft / 1000;
-      return `${Math.round(thousands * 10) / 10}K`;
-    }
-    return `${Math.round(sqft)}`;
-  }, []);
-
-  // Get slider position as percentage
-  const getSliderPosition = useCallback((sqft: number | null): number => {
-    if (sqft === null || sqft === Infinity) return 100;
-
-    const maxFiniteSqft = sqftBreakpoints[sqftBreakpoints.length - 2];
-    if (sqft >= maxFiniteSqft) return 100;
-    if (sqft <= 0) return 0;
-
-    // Find position between breakpoints
-    for (let i = 0; i < sqftBreakpoints.length - 1; i++) {
-      const current = sqftBreakpoints[i];
-      const next = sqftBreakpoints[i + 1];
-
-      if (sqft >= current && sqft <= next) {
-        const segmentSize = 100 / (sqftBreakpoints.length - 1);
-        const segmentStart = i * segmentSize;
-        const progressInSegment = next === Infinity ? 0 : (sqft - current) / (next - current);
-        return segmentStart + (progressInSegment * segmentSize);
-      }
-    }
-
-    return 0;
-  }, [sqftBreakpoints]);
-
-  // Get sqft from slider position
-  const getSqftFromPosition = useCallback((position: number): number | null => {
-    if (position >= 100) return null; // Max
-    if (position <= 0) return 0;
-
-    const segmentSize = 100 / (sqftBreakpoints.length - 1);
-    const segmentIndex = Math.floor(position / segmentSize);
-    const progressInSegment = (position % segmentSize) / segmentSize;
-
-    if (segmentIndex >= sqftBreakpoints.length - 2) return null;
-
-    const current = sqftBreakpoints[segmentIndex];
-    const next = sqftBreakpoints[segmentIndex + 1];
-
-    if (next === Infinity) return null;
-
-    return Math.round(current + (progressInSegment * (next - current)));
-  }, [sqftBreakpoints]);
-
-  // Handle mouse events for dragging
-  const handleMouseDown = useCallback((handle: 'min' | 'max') => (e: React.MouseEvent) => {
-    e.preventDefault();
-    setIsDragging(handle);
-  }, []);
-
-  const handleMouseMove = useCallback((e: MouseEvent) => {
-    if (!isDragging || !sliderRef.current) return;
-
-    const rect = sliderRef.current.getBoundingClientRect();
-    const position = Math.max(0, Math.min(100, ((e.clientX - rect.left) / rect.width) * 100));
-    const newSqft = getSqftFromPosition(position);
-
-    if (isDragging === 'min') {
-      const maxValue = tempMaxSqft === null ? Infinity : tempMaxSqft;
-      const constrainedSqft = newSqft === null ? maxValue : Math.min(newSqft, maxValue);
-      setTempMinSqft(constrainedSqft === Infinity ? 0 : constrainedSqft);
-    } else if (isDragging === 'max') {
-      const constrainedSqft = newSqft === null ? null : Math.max(newSqft, tempMinSqft);
-      setTempMaxSqft(constrainedSqft);
-    }
-  }, [isDragging, tempMinSqft, tempMaxSqft, getSqftFromPosition]);
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(null);
-  }, []);
-
-  // Add/remove global mouse event listeners
-  useEffect(() => {
-    if (isDragging) {
-      document.addEventListener('mousemove', handleMouseMove);
-      document.addEventListener('mouseup', handleMouseUp);
-      return () => {
-        document.removeEventListener('mousemove', handleMouseMove);
-        document.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [isDragging, handleMouseMove, handleMouseUp]);
-
-  // Handle keyboard events for accessibility
-  const handleKeyDown = useCallback((handle: 'min' | 'max') => (e: React.KeyboardEvent) => {
-    if (e.key === 'ArrowLeft' || e.key === 'ArrowRight') {
-      e.preventDefault();
-      const direction = e.key === 'ArrowLeft' ? -1 : 1;
-      const currentSqft = handle === 'min' ? tempMinSqft : tempMaxSqft;
-      const currentPos = getSliderPosition(currentSqft);
-      const newPos = Math.max(0, Math.min(100, currentPos + (direction * 2)));
-      const newSqft = getSqftFromPosition(newPos);
-
-      if (handle === 'min') {
-        const maxValue = tempMaxSqft === null ? Infinity : tempMaxSqft;
-        const constrainedSqft = newSqft === null ? maxValue : Math.min(newSqft, maxValue);
-        setTempMinSqft(constrainedSqft === Infinity ? 0 : constrainedSqft);
-      } else {
-        const constrainedSqft = newSqft === null ? null : Math.max(newSqft || 0, tempMinSqft);
-        setTempMaxSqft(constrainedSqft);
-      }
-    }
-  }, [tempMinSqft, tempMaxSqft, getSliderPosition, getSqftFromPosition]);
-
-  // Handle ESC key to close modal
-  useEffect(() => {
-    const handleEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' && isOpen) {
-        onCancel();
-      }
-    };
-
-    if (isOpen) {
-      document.addEventListener('keydown', handleEsc);
-      return () => document.removeEventListener('keydown', handleEsc);
-    }
-  }, [isOpen, onCancel]);
-
-  if (!isOpen) return null;
-
-  const minPosition = getSliderPosition(tempMinSqft);
-  const maxPosition = getSliderPosition(tempMaxSqft);
-
-  return (
-    <div
-      style={{
-        marginTop: "16px",
-        padding: "16px",
-        backgroundColor: "#f9fafb",
-        borderRadius: "6px",
-        border: "1px solid #e5e7eb",
-      }}
-    >
-      {/* Current Selection Display */}
-      <div style={{
-        textAlign: 'center',
-        fontSize: '14px',
-        fontWeight: '600',
-        color: '#374151',
-        marginBottom: '16px',
-      }}>
-        {formatDisplaySqft(tempMinSqft)} - {formatDisplaySqft(tempMaxSqft)} sq ft
-      </div>
-
-      {/* Slider Container */}
-      <div style={{ marginBottom: '16px' }}>
-        <div
-          ref={sliderRef}
-          style={{
-            position: 'relative',
-            height: '6px',
-            backgroundColor: '#e5e7eb',
-            borderRadius: '3px',
-            marginBottom: '12px',
-          }}
-        >
-          {/* Active Range */}
-          <div
-            style={{
-              position: 'absolute',
-              top: 0,
-              left: `${minPosition}%`,
-              width: `${maxPosition - minPosition}%`,
-              height: '100%',
-              backgroundColor: '#3b82f6',
-              borderRadius: '3px',
-            }}
-          />
-
-          {/* Min Handle */}
-          <div
-            style={{
-              position: 'absolute',
-              left: `${minPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '18px',
-              height: '18px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'min' ? 'grabbing' : 'grab',
-              boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('min')}
-            onKeyDown={handleKeyDown('min')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Minimum square footage"
-            aria-valuemin={0}
-            aria-valuemax={tempMaxSqft === null ? sqftBreakpoints[sqftBreakpoints.length - 2] : tempMaxSqft}
-            aria-valuenow={tempMinSqft}
-            aria-valuetext={formatDisplaySqft(tempMinSqft)}
-          />
-
-          {/* Max Handle */}
-          <div
-            style={{
-              position: 'absolute',
-              left: `${maxPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '18px',
-              height: '18px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'max' ? 'grabbing' : 'grab',
-              boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('max')}
-            onKeyDown={handleKeyDown('max')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Maximum square footage"
-            aria-valuemin={tempMinSqft}
-            aria-valuemax={sqftBreakpoints[sqftBreakpoints.length - 2]}
-            aria-valuenow={tempMaxSqft === null ? sqftBreakpoints[sqftBreakpoints.length - 2] : tempMaxSqft}
-            aria-valuetext={formatDisplaySqft(tempMaxSqft)}
-          />
-        </div>
-
-        {/* Sqft Markers */}
-        <div style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          fontSize: '11px',
-          color: '#6b7280',
-        }}>
-          {sqftBreakpoints.slice(0, -1).map((sqft, index) => (
-            <span key={index}>
-              {formatDisplaySqft(sqft)}
-            </span>
-          ))}
-          <span>Max</span>
-        </div>
-      </div>
-
-      {/* Action Buttons */}
-      <div style={{
-        display: 'flex',
-        gap: '8px',
-        justifyContent: 'flex-end',
-      }}>
-        <button
-          onClick={onCancel}
-          style={{
-            padding: '6px 12px',
-            fontSize: '13px',
-            fontWeight: '500',
-            color: '#374151',
-            backgroundColor: 'white',
-            border: '1px solid #d1d5db',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Cancel
-        </button>
-        <button
-          onClick={() => onApply(tempMinSqft, tempMaxSqft)}
-          style={{
-            padding: '6px 12px',
-            fontSize: '13px',
-            fontWeight: '500',
-            color: 'white',
-            backgroundColor: '#3b82f6',
-            border: 'none',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Apply
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Maintenance Fee Filter Component
-interface MaintenanceFeeFilterProps {
-  isOpen: boolean;
-  initialMin: number;
-  initialMax: number | null;
-  onApply: (min: number, max: number | null) => void;
-  onCancel: () => void;
-  feeBreakpoints?: number[];
-}
-
-function MaintenanceFeeFilter({
-  isOpen,
-  initialMin,
-  initialMax,
-  onApply,
-  onCancel,
-  feeBreakpoints = [0, 200, 400, 600, 1000, Infinity]
-}: MaintenanceFeeFilterProps) {
-  const [tempMinFee, setTempMinFee] = useState(initialMin);
-  const [tempMaxFee, setTempMaxFee] = useState(initialMax);
-  const [isDragging, setIsDragging] = useState<'min' | 'max' | null>(null);
-  const sliderRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    setTempMinFee(initialMin);
-    setTempMaxFee(initialMax);
-  }, [isOpen, initialMin, initialMax]);
-
-  const formatDisplayFee = useCallback((fee: number | null): string => {
-    if (fee === null || fee === Infinity) return "Max";
-    if (fee === 0) return "$0";
-    return `$${Math.round(fee)}`;
-  }, []);
-
-  const getSliderPosition = useCallback((fee: number | null): number => {
-    if (fee === null || fee === Infinity) return 100;
-    const maxFiniteFee = feeBreakpoints[feeBreakpoints.length - 2];
-    if (fee >= maxFiniteFee) return 100;
-    if (fee <= 0) return 0;
-
-    for (let i = 0; i < feeBreakpoints.length - 1; i++) {
-      const current = feeBreakpoints[i];
-      const next = feeBreakpoints[i + 1];
-      if (fee >= current && fee <= next) {
-        const segmentSize = 100 / (feeBreakpoints.length - 1);
-        const segmentStart = i * segmentSize;
-        const progressInSegment = next === Infinity ? 0 : (fee - current) / (next - current);
-        return segmentStart + (progressInSegment * segmentSize);
-      }
-    }
-    return 0;
-  }, [feeBreakpoints]);
-
-  const getFeeFromPosition = useCallback((position: number): number | null => {
-    if (position >= 100) return null;
-    if (position <= 0) return 0;
-
-    const segmentSize = 100 / (feeBreakpoints.length - 1);
-    const segmentIndex = Math.floor(position / segmentSize);
-    const progressInSegment = (position % segmentSize) / segmentSize;
-
-    if (segmentIndex >= feeBreakpoints.length - 2) return null;
-
-    const current = feeBreakpoints[segmentIndex];
-    const next = feeBreakpoints[segmentIndex + 1];
-
-    if (next === Infinity) return null;
-
-    return Math.round(current + (progressInSegment * (next - current)));
-  }, [feeBreakpoints]);
-
-  const handleMouseDown = useCallback((handle: 'min' | 'max') => (e: React.MouseEvent) => {
-    e.preventDefault();
-    setIsDragging(handle);
-  }, []);
-
-  const handleMouseMove = useCallback((e: MouseEvent) => {
-    if (!isDragging || !sliderRef.current) return;
-
-    const rect = sliderRef.current.getBoundingClientRect();
-    const position = Math.max(0, Math.min(100, ((e.clientX - rect.left) / rect.width) * 100));
-    const newFee = getFeeFromPosition(position);
-
-    if (isDragging === 'min') {
-      const maxValue = tempMaxFee === null ? Infinity : tempMaxFee;
-      const constrainedFee = newFee === null ? maxValue : Math.min(newFee, maxValue);
-      setTempMinFee(constrainedFee === Infinity ? 0 : constrainedFee);
-    } else if (isDragging === 'max') {
-      const constrainedFee = newFee === null ? null : Math.max(newFee, tempMinFee);
-      setTempMaxFee(constrainedFee);
-    }
-  }, [isDragging, tempMinFee, tempMaxFee, getFeeFromPosition]);
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(null);
-  }, []);
-
-  useEffect(() => {
-    if (isDragging) {
-      document.addEventListener('mousemove', handleMouseMove);
-      document.addEventListener('mouseup', handleMouseUp);
-      return () => {
-        document.removeEventListener('mousemove', handleMouseMove);
-        document.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [isDragging, handleMouseMove, handleMouseUp]);
-
-  const handleKeyDown = useCallback((handle: 'min' | 'max') => (e: React.KeyboardEvent) => {
-    if (e.key === 'ArrowLeft' || e.key === 'ArrowRight') {
-      e.preventDefault();
-      const direction = e.key === 'ArrowLeft' ? -1 : 1;
-      const currentFee = handle === 'min' ? tempMinFee : tempMaxFee;
-      const currentPos = getSliderPosition(currentFee);
-      const newPos = Math.max(0, Math.min(100, currentPos + (direction * 2)));
-      const newFee = getFeeFromPosition(newPos);
-
-      if (handle === 'min') {
-        const maxValue = tempMaxFee === null ? Infinity : tempMaxFee;
-        const constrainedFee = newFee === null ? maxValue : Math.min(newFee, maxValue);
-        setTempMinFee(constrainedFee === Infinity ? 0 : constrainedFee);
-      } else {
-        const constrainedFee = newFee === null ? null : Math.max(newFee || 0, tempMinFee);
-        setTempMaxFee(constrainedFee);
-      }
-    }
-  }, [tempMinFee, tempMaxFee, getSliderPosition, getFeeFromPosition]);
-
-  useEffect(() => {
-    const handleEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' && isOpen) {
-        onCancel();
-      }
-    };
-
-    if (isOpen) {
-      document.addEventListener('keydown', handleEsc);
-      return () => document.removeEventListener('keydown', handleEsc);
-    }
-  }, [isOpen, onCancel]);
-
-  if (!isOpen) return null;
-
-  const minPosition = getSliderPosition(tempMinFee);
-  const maxPosition = getSliderPosition(tempMaxFee);
-
-  return (
-    <div
-      style={{
-        marginTop: "16px",
-        padding: "16px",
-        backgroundColor: "#f9fafb",
-        borderRadius: "6px",
-        border: "1px solid #e5e7eb",
-      }}
-    >
-      <div style={{
-        textAlign: 'center',
-        fontSize: '14px',
-        fontWeight: '600',
-        color: '#374151',
-        marginBottom: '16px',
-      }}>
-        {formatDisplayFee(tempMinFee)} - {formatDisplayFee(tempMaxFee)} /month
-      </div>
-
-      <div style={{ marginBottom: '16px' }}>
-        <div
-          ref={sliderRef}
-          style={{
-            position: 'relative',
-            height: '6px',
-            backgroundColor: '#e5e7eb',
-            borderRadius: '3px',
-            marginBottom: '12px',
-          }}
-        >
-          <div
-            style={{
-              position: 'absolute',
-              top: 0,
-              left: `${minPosition}%`,
-              width: `${maxPosition - minPosition}%`,
-              height: '100%',
-              backgroundColor: '#3b82f6',
-              borderRadius: '3px',
-            }}
-          />
-
-          <div
-            style={{
-              position: 'absolute',
-              left: `${minPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '18px',
-              height: '18px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'min' ? 'grabbing' : 'grab',
-              boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('min')}
-            onKeyDown={handleKeyDown('min')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Minimum maintenance fee"
-            aria-valuemin={0}
-            aria-valuemax={tempMaxFee === null ? feeBreakpoints[feeBreakpoints.length - 2] : tempMaxFee}
-            aria-valuenow={tempMinFee}
-            aria-valuetext={formatDisplayFee(tempMinFee)}
-          />
-
-          <div
-            style={{
-              position: 'absolute',
-              left: `${maxPosition}%`,
-              top: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: '18px',
-              height: '18px',
-              backgroundColor: 'white',
-              border: '2px solid #3b82f6',
-              borderRadius: '50%',
-              cursor: isDragging === 'max' ? 'grabbing' : 'grab',
-              boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
-            }}
-            onMouseDown={handleMouseDown('max')}
-            onKeyDown={handleKeyDown('max')}
-            tabIndex={0}
-            role="slider"
-            aria-label="Maximum maintenance fee"
-            aria-valuemin={tempMinFee}
-            aria-valuemax={feeBreakpoints[feeBreakpoints.length - 2]}
-            aria-valuenow={tempMaxFee === null ? feeBreakpoints[feeBreakpoints.length - 2] : tempMaxFee}
-            aria-valuetext={formatDisplayFee(tempMaxFee)}
-          />
-        </div>
-
-        <div style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          fontSize: '11px',
-          color: '#6b7280',
-        }}>
-          {feeBreakpoints.slice(0, -1).map((fee, index) => (
-            <span key={index}>
-              {formatDisplayFee(fee)}
-            </span>
-          ))}
-          <span>Max</span>
-        </div>
-      </div>
-
-      <div style={{
-        display: 'flex',
-        gap: '8px',
-        justifyContent: 'flex-end',
-      }}>
-        <button
-          onClick={onCancel}
-          style={{
-            padding: '6px 12px',
-            fontSize: '13px',
-            fontWeight: '500',
-            color: '#374151',
-            backgroundColor: 'white',
-            border: '1px solid #d1d5db',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Cancel
-        </button>
-        <button
-          onClick={() => onApply(tempMinFee, tempMaxFee)}
-          style={{
-            padding: '6px 12px',
-            fontSize: '13px',
-            fontWeight: '500',
-            color: 'white',
-            backgroundColor: '#3b82f6',
-            border: 'none',
-            borderRadius: '6px',
-            cursor: 'pointer',
-          }}
-        >
-          Apply
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Filter Panel Component
-interface FilterPanelProps {
-  filters: MapFilters;
-  onFiltersChange: (filters: MapFilters) => void;
-  apiKey: string;
-}
-
-function FilterPanel({ filters, onFiltersChange, apiKey }: FilterPanelProps) {
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [isPriceFilterOpen, setIsPriceFilterOpen] = useState(false);
-  const [isActiveOpen, setIsActiveOpen] = useState(false);
-  const [isSoldOpen, setIsSoldOpen] = useState(false);
-  const [isUnavailableOpen, setIsUnavailableOpen] = useState(false);
-  const [isMoreOpen, setIsMoreOpen] = useState(false);
-  const [isSqftFilterOpen, setIsSqftFilterOpen] = useState(false);
-  const priceFilterRef = useRef<HTMLDivElement>(null);
-  const activeFilterRef = useRef<HTMLDivElement>(null);
-  const soldFilterRef = useRef<HTMLDivElement>(null);
-  const unavailableFilterRef = useRef<HTMLDivElement>(null);
-  const moreFiltersRef = useRef<HTMLDivElement>(null);
-
-  const listingTypeOptions = [
-    { value: "sale", label: "For Sale" },
-    { value: "lease", label: "For Lease" },
-  ] as const;
-
-  const handleListingTypeChange = (value: "sale" | "lease") => {
-    onFiltersChange({
-      ...filters,
-      listingType: value,
-    });
-    setIsDropdownOpen(false);
-  };
-
-  const handlePriceFilterApply = (minPrice: number, maxPrice: number | null) => {
-    onFiltersChange({
-      ...filters,
-      minPrice: minPrice || undefined,
-      maxPrice: maxPrice || undefined,
-    });
-    setIsPriceFilterOpen(false);
-  };
-
-  const handlePriceFilterCancel = () => {
-    setIsPriceFilterOpen(false);
-  };
-
-  const handleSqftFilterApply = (minSqft: number, maxSqft: number | null) => {
-    onFiltersChange({
-      ...filters,
-      minSqft: minSqft || undefined,
-      maxSqft: maxSqft || undefined,
-    });
-    setIsSqftFilterOpen(false);
-  };
-
-  const handleSqftFilterCancel = () => {
-    setIsSqftFilterOpen(false);
-  };
-
-  // Format price display for button
-  const formatPriceRange = () => {
-    const hasMinPrice = filters.minPrice && filters.minPrice > 0;
-    const hasMaxPrice = filters.maxPrice && filters.maxPrice !== null;
-
-    if (!hasMinPrice && !hasMaxPrice) return "Price";
-
-    const formatPrice = (price: number) => {
-      if (price >= 1000000) {
-        const millions = price / 1000000;
-        return `$${millions >= 10 ? Math.round(millions) : Math.round(millions * 10) / 10}M`;
-      }
-      if (price >= 1000) {
-        return `$${Math.round(price / 1000)}K`;
-      }
-      return `$${price}`;
-    };
-
-    const minText = hasMinPrice ? formatPrice(filters.minPrice!) : "$0";
-    const maxText = hasMaxPrice ? formatPrice(filters.maxPrice!) : "Max";
-
-    return `${minText} - ${maxText}`;
-  };
-
-  const selectedOption = listingTypeOptions.find(option => option.value === filters.listingType);
-
-  // Active listing filter options
-  const activeListingOptions = [
-    { value: "1" as const, label: "Last 1 days" },
-    { value: "3" as const, label: "Last 3 days" },
-    { value: "7" as const, label: "Last 7 days" },
-    { value: "30" as const, label: "Last 30 days" },
-    { value: "90" as const, label: "Last 90 days" },
-    { value: "all" as const, label: "Listing date - All" },
-    { value: "15+" as const, label: "More than 15 days" },
-    { value: "30+" as const, label: "More than 30 days" },
-    { value: "60+" as const, label: "More than 60 days" },
-    { value: "90+" as const, label: "More than 90 days" },
-  ];
-
-  // Format active filter label
-  const formatActiveLabel = () => {
-    if (filters.activeListingDays === undefined) {
-      return "Off";
-    }
-    if (filters.activeListingDays === "all") {
-      return "All";
-    }
-    const option = activeListingOptions.find(opt => opt.value === filters.activeListingDays);
-    return option ? option.label.replace("Last ", "").replace("More than ", ">") : "All";
-  };
-
-  const isActiveFilterActive = () => {
-    return filters.activeListingDays !== undefined;
-  };
-
-  // Handle active listing filter change
-  const handleActiveFilterChange = (value: typeof filters.activeListingDays) => {
-    onFiltersChange({ ...filters, activeListingDays: value, soldListingDays: undefined, unavailableListingDays: undefined });
-    setIsActiveOpen(false);
-  };
-
-  // Sold listing filter options
-  const soldListingOptions = [
-    { value: "1" as const, label: "Last 1 days" },
-    { value: "3" as const, label: "Last 3 days" },
-    { value: "7" as const, label: "Last 7 days" },
-    { value: "30" as const, label: "Last 30 days" },
-    { value: "90" as const, label: "Last 90 days" },
-    { value: "180" as const, label: "Last 180 days" },
-    { value: "360" as const, label: "Last 360 days" },
-    { value: "2025" as const, label: "Year 2025" },
-    { value: "2024" as const, label: "Year 2024" },
-    { value: "2023" as const, label: "Year 2023" },
-    { value: "2022" as const, label: "Year 2022" },
-    { value: "2021" as const, label: "Year 2021" },
-    { value: "2020" as const, label: "Year 2020" },
-    { value: "2019" as const, label: "Year 2019" },
-    { value: "2018" as const, label: "Year 2018" },
-    { value: "2017" as const, label: "Year 2017" },
-    { value: "2016" as const, label: "Year 2016" },
-    { value: "2015" as const, label: "Year 2015" },
-    { value: "2014" as const, label: "Year 2014" },
-    { value: "2013" as const, label: "Year 2013" },
-    { value: "2012" as const, label: "Year 2012" },
-    { value: "2011" as const, label: "Year 2011" },
-    { value: "2010" as const, label: "Year 2010" },
-    { value: "2009" as const, label: "Year 2009" },
-    { value: "2008" as const, label: "Year 2008" },
-    { value: "2007" as const, label: "Year 2007" },
-  ];
-
-  // Format sold filter label
-  const formatSoldLabel = () => {
-    if (!filters.soldListingDays) {
-      return "Off";
-    }
-    const option = soldListingOptions.find(opt => opt.value === filters.soldListingDays);
-    return option ? option.label.replace("Last ", "").replace("Year ", "") : "Off";
-  };
-
-  const isSoldFilterActive = () => {
-    return !!filters.soldListingDays;
-  };
-
-  // Handle sold listing filter change
-  const handleSoldFilterChange = (value: typeof filters.soldListingDays) => {
-    onFiltersChange({ ...filters, soldListingDays: value, activeListingDays: undefined, unavailableListingDays: undefined });
-    setIsSoldOpen(false);
-  };
-
-  // Unavailable listing filter options (same as sold)
-  const unavailableListingOptions = [
-    { value: "1" as const, label: "Last 1 days" },
-    { value: "3" as const, label: "Last 3 days" },
-    { value: "7" as const, label: "Last 7 days" },
-    { value: "30" as const, label: "Last 30 days" },
-    { value: "90" as const, label: "Last 90 days" },
-    { value: "180" as const, label: "Last 180 days" },
-    { value: "360" as const, label: "Last 360 days" },
-    { value: "2025" as const, label: "Year 2025" },
-    { value: "2024" as const, label: "Year 2024" },
-    { value: "2023" as const, label: "Year 2023" },
-    { value: "2022" as const, label: "Year 2022" },
-    { value: "2021" as const, label: "Year 2021" },
-    { value: "2020" as const, label: "Year 2020" },
-    { value: "2019" as const, label: "Year 2019" },
-    { value: "2018" as const, label: "Year 2018" },
-    { value: "2017" as const, label: "Year 2017" },
-    { value: "2016" as const, label: "Year 2016" },
-    { value: "2015" as const, label: "Year 2015" },
-    { value: "2014" as const, label: "Year 2014" },
-    { value: "2013" as const, label: "Year 2013" },
-    { value: "2012" as const, label: "Year 2012" },
-    { value: "2011" as const, label: "Year 2011" },
-    { value: "2010" as const, label: "Year 2010" },
-    { value: "2009" as const, label: "Year 2009" },
-    { value: "2008" as const, label: "Year 2008" },
-    { value: "2007" as const, label: "Year 2007" },
-  ];
-
-  // Format unavailable filter label
-  const formatUnavailableLabel = () => {
-    if (!filters.unavailableListingDays) {
-      return "Off";
-    }
-    const option = unavailableListingOptions.find(opt => opt.value === filters.unavailableListingDays);
-    return option ? option.label.replace("Last ", "").replace("Year ", "") : "Off";
-  };
-
-  const isUnavailableFilterActive = () => {
-    return !!filters.unavailableListingDays;
-  };
-
-  // Handle unavailable listing filter change
-  const handleUnavailableFilterChange = (value: typeof filters.unavailableListingDays) => {
-    onFiltersChange({ ...filters, unavailableListingDays: value, activeListingDays: undefined, soldListingDays: undefined });
-    setIsUnavailableOpen(false);
-  };
-
-  // Handle click outside to close price filter dropdown
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (priceFilterRef.current && !priceFilterRef.current.contains(event.target as Node)) {
-        setIsPriceFilterOpen(false);
-      }
-    };
-
-    if (isPriceFilterOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-      return () => document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [isPriceFilterOpen]);
-
-  // Handle click outside to close active filter dropdown
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (activeFilterRef.current && !activeFilterRef.current.contains(event.target as Node)) {
-        setIsActiveOpen(false);
-      }
-    };
-
-    if (isActiveOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-      return () => document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [isActiveOpen]);
-
-  // Handle click outside to close sold filter dropdown
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (soldFilterRef.current && !soldFilterRef.current.contains(event.target as Node)) {
-        setIsSoldOpen(false);
-      }
-    };
-
-    if (isSoldOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-      return () => document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [isSoldOpen]);
-
-  // Handle click outside to close unavailable filter dropdown
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (unavailableFilterRef.current && !unavailableFilterRef.current.contains(event.target as Node)) {
-        setIsUnavailableOpen(false);
-      }
-    };
-
-    if (isUnavailableOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-      return () => document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [isUnavailableOpen]);
-
-  // Handle click outside to close more filters dropdown
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (moreFiltersRef.current && !moreFiltersRef.current.contains(event.target as Node)) {
-        setIsMoreOpen(false);
-      }
-    };
-
-    if (isMoreOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-      return () => document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [isMoreOpen]);
-
-  const handleClearFilters = () => {
-    onFiltersChange({
-      listingType: "sale",
-      propertyTypes: [],
-      minPrice: undefined,
-      maxPrice: null,
-      bedrooms: undefined,
-      bathrooms: undefined,
-      garageSpaces: undefined,
-      minSqft: undefined,
-      maxSqft: null,
-      openHouse: undefined,
-      maxMaintenanceFee: null,
-      activeListingDays: "all",
-      soldListingDays: undefined,
-      unavailableListingDays: undefined,
-    });
-  };
-
-  return (
-    <div
-      style={{
-        position: "absolute",
-        top: "16px",
-        left: "16px",
-        backgroundColor: "rgba(255, 255, 255, 0.95)",
-        borderRadius: "8px",
-        boxShadow: "0 2px 8px rgba(0, 0, 0, 0.15)",
-        padding: "12px",
-        fontSize: "14px",
-        fontWeight: "500",
-        color: "#374151",
-        zIndex: 1000,
-        minWidth: "200px",
-      }}
-    >
-      <div style={{
-        marginBottom: "8px",
-        fontSize: "12px",
-        fontWeight: "600",
-        color: "#6b7280",
-        display: "flex",
-        justifyContent: "space-between",
-        alignItems: "center",
-      }}>
-        <span>FILTERS</span>
-        <button
-          onClick={handleClearFilters}
-          style={{
-            padding: "2px 8px",
-            fontSize: "11px",
-            fontWeight: "500",
-            color: "#6b7280",
-            backgroundColor: "transparent",
-            border: "1px solid #d1d5db",
-            borderRadius: "4px",
-            cursor: "pointer",
-          }}
-        >
-          Clear
-        </button>
-      </div>
-
-      {/* Listing Type Dropdown */}
-      <div style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsDropdownOpen(!isDropdownOpen)}
-          style={{
-            width: "100%",
-            padding: "8px 12px",
-            backgroundColor: filters.listingType === "sale" ? "#d1fae5" : "#f3e8ff",
-            border: filters.listingType === "sale" ? "1px solid #86efac" : "1px solid #d8b4fe",
-            borderRadius: "6px",
-            fontSize: "14px",
-            fontWeight: "500",
-            cursor: "pointer",
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-            color: "#1f2937",
-          }}
-        >
-          <span>{selectedOption?.label}</span>
-          <ChevronDown
-            size={16}
-            style={{
-              transform: isDropdownOpen ? "rotate(180deg)" : "rotate(0deg)",
-              transition: "transform 0.2s"
-            }}
-          />
-        </button>
-
-        {isDropdownOpen && (
-          <div
-            style={{
-              position: "absolute",
-              top: "100%",
-              left: 0,
-              right: 0,
-              backgroundColor: "white",
-              border: "1px solid #d1d5db",
-              borderRadius: "6px",
-              marginTop: "4px",
-              boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-              zIndex: 1001,
-            }}
-          >
-            {listingTypeOptions.map((option) => (
-              <button
-                key={option.value}
-                onClick={() => handleListingTypeChange(option.value)}
-                style={{
-                  width: "100%",
-                  padding: "8px 12px",
-                  backgroundColor: filters.listingType === option.value ? "#f3f4f6" : "white",
-                  border: "none",
-                  fontSize: "14px",
-                  cursor: "pointer",
-                  textAlign: "left",
-                  borderRadius: option === listingTypeOptions[0] ? "6px 6px 0 0" :
-                            option === listingTypeOptions[listingTypeOptions.length - 1] ? "0 0 6px 6px" : "0",
-                }}
-                onMouseEnter={(e) => {
-                  if (filters.listingType !== option.value) {
-                    e.currentTarget.style.backgroundColor = "#f9fafb";
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (filters.listingType !== option.value) {
-                    e.currentTarget.style.backgroundColor = "white";
-                  }
-                }}
-              >
-                {option.label}
-              </button>
-            ))}
-          </div>
-        )}
-      </div>
-
-      {/* Property Type Filter */}
-      <PropertyTypeFilter
-        filters={filters}
-        onFiltersChange={onFiltersChange}
-        apiKey={apiKey}
-      />
-
-      {/* Price Filter Button */}
-      <div ref={priceFilterRef} style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsPriceFilterOpen(!isPriceFilterOpen)}
-          style={{
-            width: "100%",
-            padding: "8px 12px",
-            backgroundColor: (filters.minPrice || filters.maxPrice) ? "#f3f4f6" : "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            fontSize: "14px",
-            cursor: "pointer",
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-            textAlign: "left",
-            color: (filters.minPrice || filters.maxPrice) ? "#1f2937" : "#374151",
-            fontWeight: (filters.minPrice || filters.maxPrice) ? "500" : "400",
-          }}
-        >
-          <span>{formatPriceRange()}</span>
-          <ChevronDown
-            size={16}
-            style={{
-              transform: isPriceFilterOpen ? "rotate(180deg)" : "rotate(0deg)",
-              transition: "transform 0.2s"
-            }}
-          />
-        </button>
-
-        {/* Price Range Filter Dropdown */}
-        {isPriceFilterOpen && (
-          <PriceRangeFilter
-            isOpen={isPriceFilterOpen}
-            initialMin={filters.minPrice || 0}
-            initialMax={filters.maxPrice || null}
-            onApply={handlePriceFilterApply}
-            onCancel={handlePriceFilterCancel}
-          />
-        )}
-      </div>
-
-      {/* Active Listings Filter Button */}
-      <div ref={activeFilterRef} style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsActiveOpen(!isActiveOpen)}
-          style={{
-            width: "100%",
-            padding: "0",
-            backgroundColor: "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            fontSize: "14px",
-            cursor: "pointer",
-            display: "flex",
-            overflow: "hidden",
-          }}
-        >
-          <div
-            style={{
-              width: "48px",
-              backgroundColor: isActiveFilterActive() ? "#10b981" : "#e5e7eb",
-              flexShrink: 0,
-            }}
-          />
-          <div
-            style={{
-              flex: 1,
-              padding: "8px 12px",
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              textAlign: "left",
-              color: "#1f2937",
-              fontWeight: "500",
-            }}
-          >
-            <span>Active: {formatActiveLabel()}</span>
-            <ChevronDown
-              size={16}
-              style={{
-                transform: isActiveOpen ? "rotate(180deg)" : "rotate(0deg)",
-                transition: "transform 0.2s"
-              }}
-            />
-          </div>
-        </button>
-
-        {/* Active Filter Dropdown */}
-        {isActiveOpen && (
-          <div
-            style={{
-              position: "absolute",
-              top: "100%",
-              left: 0,
-              right: 0,
-              backgroundColor: "white",
-              border: "1px solid #d1d5db",
-              borderRadius: "6px",
-              marginTop: "4px",
-              padding: "12px",
-              boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-              zIndex: 1001,
-            }}
-          >
-            {activeListingOptions.map((option) => (
-              <label
-                key={option.value}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  padding: "8px",
-                  cursor: "pointer",
-                  borderRadius: "4px",
-                  transition: "background-color 0.2s",
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = "#f9fafb";
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = "transparent";
-                }}
-              >
-                <input
-                  type="radio"
-                  name="activeListingFilter"
-                  checked={filters.activeListingDays === option.value || (!filters.activeListingDays && option.value === "all")}
-                  onChange={() => handleActiveFilterChange(option.value)}
-                  style={{
-                    marginRight: "8px",
-                    width: "16px",
-                    height: "16px",
-                    cursor: "pointer",
-                  }}
-                />
-                <span style={{ fontSize: "14px", color: "#374151" }}>{option.label}</span>
-              </label>
-            ))}
-          </div>
-        )}
-      </div>
-
-      {/* Sold Listings Filter Button */}
-      <div ref={soldFilterRef} style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsSoldOpen(!isSoldOpen)}
-          style={{
-            width: "100%",
-            padding: "0",
-            backgroundColor: "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            fontSize: "14px",
-            cursor: "pointer",
-            display: "flex",
-            overflow: "hidden",
-          }}
-        >
-          <div
-            style={{
-              width: "48px",
-              backgroundColor: isSoldFilterActive() ? "#8b7fa8" : "#e5e7eb",
-              flexShrink: 0,
-            }}
-          />
-          <div
-            style={{
-              flex: 1,
-              padding: "8px 12px",
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              textAlign: "left",
-              color: "#1f2937",
-              fontWeight: "500",
-            }}
-          >
-            <span>Sold: {formatSoldLabel()}</span>
-            <ChevronDown
-              size={16}
-              style={{
-                transform: isSoldOpen ? "rotate(180deg)" : "rotate(0deg)",
-                transition: "transform 0.2s"
-              }}
-            />
-          </div>
-        </button>
-
-        {/* Sold Filter Dropdown */}
-        {isSoldOpen && (
-          <div
-            style={{
-              position: "absolute",
-              top: "100%",
-              left: 0,
-              right: 0,
-              backgroundColor: "white",
-              border: "1px solid #d1d5db",
-              borderRadius: "6px",
-              marginTop: "4px",
-              padding: "12px",
-              boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-              zIndex: 1001,
-              maxHeight: "400px",
-              overflowY: "auto",
-            }}
-          >
-            {soldListingOptions.map((option) => (
-              <label
-                key={option.value}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  padding: "8px",
-                  cursor: "pointer",
-                  borderRadius: "4px",
-                  transition: "background-color 0.2s",
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = "#f9fafb";
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = "transparent";
-                }}
-              >
-                <input
-                  type="radio"
-                  name="soldListingFilter"
-                  checked={filters.soldListingDays === option.value}
-                  onChange={() => handleSoldFilterChange(option.value)}
-                  style={{
-                    marginRight: "8px",
-                    width: "16px",
-                    height: "16px",
-                    cursor: "pointer",
-                  }}
-                />
-                <span style={{ fontSize: "14px", color: "#374151" }}>{option.label}</span>
-              </label>
-            ))}
-          </div>
-        )}
-      </div>
-
-      {/* Unavailable Listings Filter Button */}
-      <div ref={unavailableFilterRef} style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsUnavailableOpen(!isUnavailableOpen)}
-          style={{
-            width: "100%",
-            padding: "0",
-            backgroundColor: "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            fontSize: "14px",
-            cursor: "pointer",
-            display: "flex",
-            overflow: "hidden",
-          }}
-        >
-          <div
-            style={{
-              width: "48px",
-              backgroundColor: isUnavailableFilterActive() ? "#f59e0b" : "#e5e7eb",
-              flexShrink: 0,
-            }}
-          />
-          <div
-            style={{
-              flex: 1,
-              padding: "8px 12px",
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              textAlign: "left",
-              color: "#1f2937",
-              fontWeight: "500",
-            }}
-          >
-            <span>Unavailable: {formatUnavailableLabel()}</span>
-            <ChevronDown
-              size={16}
-              style={{
-                transform: isUnavailableOpen ? "rotate(180deg)" : "rotate(0deg)",
-                transition: "transform 0.2s"
-              }}
-            />
-          </div>
-        </button>
-
-        {/* Unavailable Filter Dropdown */}
-        {isUnavailableOpen && (
-          <div
-            style={{
-              position: "absolute",
-              top: "100%",
-              left: 0,
-              right: 0,
-              backgroundColor: "white",
-              border: "1px solid #d1d5db",
-              borderRadius: "6px",
-              marginTop: "4px",
-              padding: "12px",
-              boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-              zIndex: 1001,
-              maxHeight: "400px",
-              overflowY: "auto",
-            }}
-          >
-            {unavailableListingOptions.map((option) => (
-              <label
-                key={option.value}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  padding: "8px",
-                  cursor: "pointer",
-                  borderRadius: "4px",
-                  transition: "background-color 0.2s",
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = "#f9fafb";
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = "transparent";
-                }}
-              >
-                <input
-                  type="radio"
-                  name="unavailableListingFilter"
-                  checked={filters.unavailableListingDays === option.value}
-                  onChange={() => handleUnavailableFilterChange(option.value)}
-                  style={{
-                    marginRight: "8px",
-                    width: "16px",
-                    height: "16px",
-                    cursor: "pointer",
-                  }}
-                />
-                <span style={{ fontSize: "14px", color: "#374151" }}>{option.label}</span>
-              </label>
-            ))}
-          </div>
-        )}
-      </div>
-
-      {/* More Filters Button */}
-      <div ref={moreFiltersRef} style={{ position: "relative", marginBottom: "12px" }}>
-        <button
-          onClick={() => setIsMoreOpen(!isMoreOpen)}
-          style={{
-            width: "100%",
-            padding: "8px 12px",
-            backgroundColor: (filters.bedrooms !== "all" || filters.bathrooms !== "all" || filters.garageSpaces !== "all" || filters.minSqft || filters.maxSqft || (filters.openHouse && filters.openHouse !== "all") || filters.maxMaintenanceFee) ? "#f3f4f6" : "white",
-            border: "1px solid #d1d5db",
-            borderRadius: "6px",
-            fontSize: "14px",
-            cursor: "pointer",
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-            textAlign: "left",
-            color: "#374151",
-            fontWeight: (filters.bedrooms !== "all" || filters.bathrooms !== "all" || filters.garageSpaces !== "all" || filters.minSqft || filters.maxSqft || (filters.openHouse && filters.openHouse !== "all") || filters.maxMaintenanceFee) ? "500" : "400",
-          }}
-        >
-          <span>More Filters</span>
-          <ChevronDown
-            size={16}
-            style={{
-              transform: isMoreOpen ? "rotate(180deg)" : "rotate(0deg)",
-              transition: "transform 0.2s"
-            }}
-          />
-        </button>
-
-        {/* More Filters Dropdown */}
-        {isMoreOpen && (
-          <div
-            style={{
-              position: "absolute",
-              top: "100%",
-              left: 0,
-              width: "400px",
-              backgroundColor: "white",
-              border: "1px solid #d1d5db",
-              borderRadius: "6px",
-              marginTop: "4px",
-              padding: "20px",
-              boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-              zIndex: 1001,
-            }}
-          >
-            {/* Bedrooms Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                BEDROOMS
-              </div>
-              <div style={{ display: "flex", gap: "8px", flexWrap: "wrap" }}>
-                {["all", "0", "1", "2", "3", "4", "5+"].map((value) => (
-                  <button
-                    key={value}
-                    onClick={() => onFiltersChange({ ...filters, bedrooms: value as any })}
-                    style={{
-                      padding: "8px 16px",
-                      backgroundColor: filters.bedrooms === value ? "#3b82f6" : "white",
-                      color: filters.bedrooms === value ? "white" : "#374151",
-                      border: `1px solid ${filters.bedrooms === value ? "#3b82f6" : "#d1d5db"}`,
-                      borderRadius: "6px",
-                      fontSize: "14px",
-                      fontWeight: filters.bedrooms === value ? "600" : "400",
-                      cursor: "pointer",
-                      transition: "all 0.2s",
-                    }}
-                    onMouseEnter={(e) => {
-                      if (filters.bedrooms !== value) {
-                        e.currentTarget.style.backgroundColor = "#f3f4f6";
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (filters.bedrooms !== value) {
-                        e.currentTarget.style.backgroundColor = "white";
-                      }
-                    }}
-                  >
-                    {value === "all" ? "All" : value}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            {/* Bathrooms Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                BATHROOMS
-              </div>
-              <div style={{ display: "flex", gap: "8px", flexWrap: "wrap" }}>
-                {["all", "1+", "2+", "3+", "4+", "5+"].map((value) => (
-                  <button
-                    key={value}
-                    onClick={() => onFiltersChange({ ...filters, bathrooms: value as any })}
-                    style={{
-                      padding: "8px 16px",
-                      backgroundColor: filters.bathrooms === value ? "#3b82f6" : "white",
-                      color: filters.bathrooms === value ? "white" : "#374151",
-                      border: `1px solid ${filters.bathrooms === value ? "#3b82f6" : "#d1d5db"}`,
-                      borderRadius: "6px",
-                      fontSize: "14px",
-                      fontWeight: filters.bathrooms === value ? "600" : "400",
-                      cursor: "pointer",
-                      transition: "all 0.2s",
-                    }}
-                    onMouseEnter={(e) => {
-                      if (filters.bathrooms !== value) {
-                        e.currentTarget.style.backgroundColor = "#f3f4f6";
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (filters.bathrooms !== value) {
-                        e.currentTarget.style.backgroundColor = "white";
-                      }
-                    }}
-                  >
-                    {value === "all" ? "All" : value}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            {/* Garage/Parking Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                GARAGE / PARKING
-              </div>
-              <div style={{ display: "flex", gap: "8px", flexWrap: "wrap" }}>
-                {["all", "1+", "2+", "3+", "4+", "5+"].map((value) => (
-                  <button
-                    key={value}
-                    onClick={() => onFiltersChange({ ...filters, garageSpaces: value as any })}
-                    style={{
-                      padding: "8px 16px",
-                      backgroundColor: filters.garageSpaces === value ? "#3b82f6" : "white",
-                      color: filters.garageSpaces === value ? "white" : "#374151",
-                      border: `1px solid ${filters.garageSpaces === value ? "#3b82f6" : "#d1d5db"}`,
-                      borderRadius: "6px",
-                      fontSize: "14px",
-                      fontWeight: filters.garageSpaces === value ? "600" : "400",
-                      cursor: "pointer",
-                      transition: "all 0.2s",
-                    }}
-                    onMouseEnter={(e) => {
-                      if (filters.garageSpaces !== value) {
-                        e.currentTarget.style.backgroundColor = "#f3f4f6";
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (filters.garageSpaces !== value) {
-                        e.currentTarget.style.backgroundColor = "white";
-                      }
-                    }}
-                  >
-                    {value === "all" ? "All" : value}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            {/* Square Footage Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                SQUARE FOOTAGE
-              </div>
-              <SquareFootageFilter
-                isOpen={true}
-                initialMin={filters.minSqft || 0}
-                initialMax={filters.maxSqft || null}
-                onApply={handleSqftFilterApply}
-                onCancel={handleSqftFilterCancel}
-              />
-            </div>
-
-            {/* Open House Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                OPEN HOUSE
-              </div>
-              <div style={{ display: "flex", gap: "8px", flexWrap: "wrap" }}>
-                {[
-                  { value: "all", label: "All" },
-                  { value: "today", label: "Today" },
-                  { value: "thisWeekend", label: "This Weekend" },
-                  { value: "thisWeek", label: "This Week" },
-                  { value: "anytime", label: "Anytime" }
-                ].map(({ value, label }) => (
-                  <button
-                    key={value}
-                    onClick={() => onFiltersChange({ ...filters, openHouse: value as any })}
-                    style={{
-                      padding: "8px 16px",
-                      backgroundColor: filters.openHouse === value ? "#3b82f6" : "white",
-                      color: filters.openHouse === value ? "white" : "#374151",
-                      border: `1px solid ${filters.openHouse === value ? "#3b82f6" : "#d1d5db"}`,
-                      borderRadius: "6px",
-                      fontSize: "14px",
-                      fontWeight: filters.openHouse === value ? "600" : "400",
-                      cursor: "pointer",
-                      transition: "all 0.2s",
-                    }}
-                    onMouseEnter={(e) => {
-                      if (filters.openHouse !== value) {
-                        e.currentTarget.style.backgroundColor = "#f3f4f6";
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (filters.openHouse !== value) {
-                        e.currentTarget.style.backgroundColor = "white";
-                      }
-                    }}
-                  >
-                    {label}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            {/* Maintenance Fee Filter */}
-            <div>
-              <div style={{ fontSize: "12px", fontWeight: "600", color: "#6b7280", marginBottom: "8px" }}>
-                MAX MAINTENANCE FEE
-              </div>
-              <div style={{ display: "flex", gap: "8px", alignItems: "center" }}>
-                <input
-                  type="number"
-                  placeholder="Enter max fee"
-                  value={filters.maxMaintenanceFee || ""}
-                  onChange={(e) => {
-                    const value = e.target.value ? parseInt(e.target.value) : undefined;
-                    onFiltersChange({ ...filters, maxMaintenanceFee: value || null });
-                  }}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                      const value = e.currentTarget.value ? parseInt(e.currentTarget.value) : undefined;
-                      onFiltersChange({ ...filters, maxMaintenanceFee: value || null });
-                    }
-                  }}
-                  style={{
-                    flex: 1,
-                    padding: "8px 12px",
-                    border: "1px solid #d1d5db",
-                    borderRadius: "6px",
-                    fontSize: "14px",
-                    color: "#374151",
-                  }}
-                />
-                <button
-                  onClick={() => {
-                    // Just trigger a re-render to apply the current value
-                    onFiltersChange({ ...filters });
-                  }}
-                  style={{
-                    padding: "8px 16px",
-                    fontSize: "14px",
-                    fontWeight: "500",
-                    color: "white",
-                    backgroundColor: "#3b82f6",
-                    border: "none",
-                    borderRadius: "6px",
-                    cursor: "pointer",
-                    whiteSpace: "nowrap",
-                  }}
-                >
-                  Apply
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-/**
- * MapListings Component
- *
- * A high-performance real estate map component that displays property listings
- * using server-side clustering from the Repliers API. Automatically adjusts
- * cluster precision based on zoom level for optimal user experience.
- *
- * Features:
- * - Server-side clustering for optimal performance
- * - Zoom-based precision adjustment
- * - Hierarchical cluster drilling
- * - Minimal API calls
- * - Property count display
- * - Filter panel for search refinement
- */
 export function MapListings({
   apiKey,
   mapboxToken,
@@ -3445,65 +84,6 @@ export function MapListings({
   const handleFiltersChange = useCallback((newFilters: MapFilters) => {
     setFilters(newFilters);
   }, []);
-
-  // Format price based on property type
-  const formatPrice = useCallback((price: number, type: string): string => {
-    const formatToMaxDigits = (value: number, suffix: string): string => {
-      if (value >= 100) {
-        // 3 digits: 189K, 999K, 234M
-        return `${Math.round(value)}${suffix}`;
-      } else if (value >= 10) {
-        // 2 digits + 1 decimal: 39.9K, 12.5M
-        return `${Math.round(value * 10) / 10}${suffix}`;
-      } else {
-        // 1 digit + 2 decimals: 6.56K, 1.45M
-        return `${Math.round(value * 100) / 100}${suffix}`;
-      }
-    };
-
-    if (type === "Lease") {
-      // Lease properties - always show as K
-      if (price >= 1000) {
-        const thousands = price / 1000;
-        return formatToMaxDigits(thousands, 'K');
-      }
-      return Math.round(price).toString(); // Round small values to whole numbers
-    } else {
-      // Sale properties - prefer M over K when >= 1M, and handle edge cases better
-      if (price >= 999500) { // Treat 999.5K+ as 1M for better UX
-        const millions = price / 1000000;
-        return formatToMaxDigits(millions, 'M');
-      } else if (price >= 1000) {
-        const thousands = price / 1000;
-        return formatToMaxDigits(thousands, 'K');
-      }
-      return Math.round(price).toString(); // Round small values to whole numbers
-    }
-  }, []);
-
-  // Create price bubble HTML element (back to simple version)
-  const createPriceBubble = useCallback((price: number, type: string): HTMLElement => {
-    const formattedPrice = formatPrice(price, type);
-    const isLease = type === "Lease";
-
-    const bubble = document.createElement('div');
-    bubble.style.cssText = `
-      background-color: ${isLease ? '#a855f7' : '#22c55e'};
-      color: white;
-      padding: 4px 8px;
-      border-radius: 4px;
-      font-size: 11px;
-      font-weight: 600;
-      font-family: -apple-system, BlinkMacSystemFont, sans-serif;
-      box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-      white-space: nowrap;
-      pointer-events: auto;
-      cursor: pointer;
-    `;
-    bubble.textContent = formattedPrice;
-
-    return bubble;
-  }, [formatPrice]);
 
   // Clear existing price markers
   const clearPriceMarkers = useCallback(() => {
@@ -3634,7 +214,7 @@ export function MapListings({
         priceMarkers.current.push(marker);
       }
     });
-  }, [clearPriceMarkers, createPriceBubble, apiKey]);
+  }, [clearPriceMarkers, apiKey]);
 
   // Auto-center detection effect
   useEffect(() => {
@@ -4393,37 +973,6 @@ export function MapListings({
 
     return [...largeClusters, ...mergedFeatures];
   }, [groupFeaturesByProximity]);
-
-  // Find individual listings near a cluster coordinate for tooltip display
-  const findNearbyListings = useCallback((clusterCoords: [number, number], radiusMeters: number, maxCount: number): any[] => {
-    if (!map.current) return [];
-
-    try {
-      const source = map.current.getSource('listings') as mapboxgl.GeoJSONSource;
-      if (!source || !source._data) return [];
-
-      const sourceData = source._data as any;
-      const allFeatures = sourceData.features || [];
-
-      // Find property features (not clusters) near the cluster
-      const nearbyProperties = allFeatures
-        .filter((feature: any) => feature.properties?.isProperty === true)
-        .map((feature: any) => {
-          const featureCoords = feature.geometry.coordinates as [number, number];
-          const distance = calculateDistance(clusterCoords, featureCoords);
-          return { feature, distance, listing: feature.properties };
-        })
-        .filter((item: any) => item.distance <= radiusMeters)
-        .sort((a: any, b: any) => a.distance - b.distance)
-        .slice(0, maxCount)
-        .map((item: any) => item.listing);
-
-      return nearbyProperties;
-
-    } catch (error) {
-      return [];
-    }
-  }, [calculateDistance]);
 
   // Initialize map
   useEffect(() => {
